% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_diamond_ratio.R
\name{CI_diamond_ratio}
\alias{CI_diamond_ratio}
\title{Estimate diamond ratio}
\usage{
CI_diamond_ratio(RE, FE, vi, conf_level = 0.95)
}
\arguments{
\item{RE}{metafor object with random effects result}

\item{FE}{metafor object with fixed effects result}

\item{vi}{vector of effect size variances}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject the DR
}
\description{
\code{CI_diamond_ratio} returns
}
