% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareindmeanssummary.h.R
\name{compareindmeanssummary}
\alias{compareindmeanssummary}
\title{Compare Two Means - Summary Data}
\usage{
compareindmeanssummary(m1, m2, s1, s2, n1, n2, conf.level = 0.95,
  var.equal = TRUE, ymin = "auto", ymax = "auto", ylab = "auto",
  xlab = "auto", g1lab = "Comparison Group",
  g2lab = "Reference Group", g1color = "white", g2color = "gray75",
  bsize = 2, ropeBottom = "none", ropeTop = "none")
}
\arguments{
\item{m1}{.}

\item{m2}{.}

\item{s1}{.}

\item{s2}{.}

\item{n1}{.}

\item{n2}{.}

\item{conf.level}{.}

\item{var.equal}{.}

\item{ymin}{.}

\item{ymax}{.}

\item{ylab}{.}

\item{xlab}{.}

\item{g1lab}{.}

\item{g2lab}{.}

\item{g1color}{.}

\item{g2color}{.}

\item{bsize}{.}

\item{ropeBottom}{.}

\item{ropeTop}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$ctmTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$notes} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$smd} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$decisionMakingText} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$decisionMaking} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$ctmTable$asDF}

\code{as.data.frame(results$ctmTable)}
}
\description{
Compare Two Means - Summary Data
}
