% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_basol_badnews}
\alias{data_basol_badnews}
\title{Basol badnews - Ch07 - from Basol et al. (2020)}
\format{
\subsection{\code{data_basol_badnews}}{

A data frame with 198 rows and 3 columns:
\describe{
\item{Diff reliability}{numeric}
\item{Diff confidence}{numeric}
\item{Condition}{factor}
}
}
}
\source{
\doi{10.5334/joc.91}
}
\usage{
data_basol_badnews
}
\description{
Climate change? Vaccines? Fake news and conspiracy theories on these and
numerous other issues can be highly damaging, but are thriving in this social
media age. Trying to debunk a conspiracy theory by presenting facts and
evidence often doesn't work, alas. Psychological inoculation, also similar to
prebunking, presents a mild form of misinformation, preferably with
explanation, in the hope of building resistance to real-life fake news-a sort
of vaccine for fake news. The Bad News game is a spin-off from research on
psychological inoculation. Basol et al. (2020) assessed the possible
effectiveness of this game as a fake news vaccine. At getbadnews.com you can
click 'About' for information, or just start playing the game- it's easy and
maybe even fun. You encounter mock Twitter (now X) fake news messages that
illustrate common strategies for making fake news memorable or believable.
You make choices between messages and decide which ones to 'forward' as you
try to spread fake news while building your credibility score and number of
'followers'-rather like real life for a conspiracy theorist wanting to spread
the word. Compete with your friends for credibility and number of followers.
Basol's online participants first saw 18 fictitious fake news tweets and
rated each for reliability (accuracy, believability), and also rated their
confidence in that reliability rating. Both ratings were on a 1 to 7 scale.
Those in the BadNews group then played the game for 15 minutes, whereas those
in the Control group played Tetris. Then all once again gave reliability and
confidence ratings for the 18 tweets.
}
\keyword{datasets}
