% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bem_and_honorton_1994}
\alias{data_bem_and_honorton_1994}
\title{Sample data for estimate_proportion: Ten Ganzfeld studies from Bem and
Honorton (1994)}
\format{
\subsection{\code{data_bem_and_honorton_1994} A data frame with 10 rows and 3}{

columns:
\describe{
\item{study}{Name of the study}
\item{hits}{Number of correct responses}
\item{trials}{Number of trials}
}
}
}
\source{
Bem, D. J., & Honorton, C. (1994). Does Psi exist? Replicable
evidence for an anomalous process of information transfer.
\emph{Psychological Bulletin}, 115, 4–-18. doi:10.1037/0033-2909.115.1.4
}
\usage{
data_bem_and_honorton_1994
}
\description{
Three pilot studies helped refine the procedures, then four studies used
novice receivers. Study 5 used 20 students of music, drama, or dance as
receivers, in response to suggestions that creative people might be more
likely to show telepathy. Studies 6 and 7 used receivers who had participated
in an earlier study. The proportion of hits expected by chance is .25
}
\keyword{datasets}
