% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bem_psychic}
\alias{data_bem_psychic}
\title{Bem Psychic - Ch13 - from Bem and Honorton (1994)}
\format{
\subsection{\code{data_bem_psychic}}{

A data frame with 10 rows and 5 columns:
\describe{
\item{Study}{factor}
\item{Participants}{factor}
\item{N(Trials)}{integer}
\item{N(Hits)}{integer}
\item{Proportion Hits}{numeric}
}
}
}
\source{
\doi{10.1037/0033-2909.115.1.25}
}
\usage{
data_bem_psychic
}
\description{
Daryl Bem was an experienced mentalist and research psychologist, who, a
decade earlier, had been one of several outside experts invited to scrutinize
the laboratory and experimental procedures of parapsychology researcher
Charles Honorton. Bem not only judged them adequate, but joined the research
effort and became a coauthor with Honorton. Bem and Honorton (1994) first
reviewed early ganzfeld studies and described how the experimental procedure
had been improved to reduce the chance that results could be influenced by
various possible biases, or leakages of information from sender to receiver.
For example, the randomization procedure was carried out automatically by
computer, and all stimuli were presented under computer control. Bem and
Honorton then presented data from studies conducted with the improved
procedure. Table 13.1 presents basic data from 10 studies reported by Bem and
Honorton (1994). Participants each made a single judgment, so in Pilot 1, for
example, 22 participants responded, with 8 of them giving a correct response.
Three pilot studies helped refine the procedures, then four studies used
novice receivers. Study 5 used 20 students of music, drama, or dance as
receivers, in response to suggestions that creative people might be more
likely to show telepathy. Studies 6 and 7 used receivers who had participated
in an earlier study. The proportion of hits expected by chance is .25, and
Table 13.1 shows that all but Study 1 found proportions higher than .25.
}
\keyword{datasets}
