% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bodywellf}
\alias{data_bodywellf}
\title{BodyWellF - Ch12 - Body Satisfaction and Well-being data for females from Figure 11.24 right panel}
\format{
\subsection{\code{data_bodywellf}}{

A data frame with 59 rows and 2 columns:
\describe{
\item{Body Satisfaction}{numeric}
\item{Well-being}{numeric}
}
}
}
\usage{
data_bodywellf
}
\description{
A subset of data_bodywell_fm, reports only for those participants
who identified as female.  Data is Subjective Wellbeing and Body Satisfaction.
}
\keyword{datasets}
