% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bodywellfm}
\alias{data_bodywellfm}
\title{BodyWellFM - Ch12 - Body Satisfaction and Well-being data from Figure 11.1}
\format{
\subsection{\code{data_bodywellfm}}{

A data frame with 106 rows and 2 columns:
\describe{
\item{Body Satisfaction}{numeric}
\item{Well-being}{numeric}
}
}
}
\usage{
data_bodywellfm
}
\description{
Survey data from a convenience sample Dominican University students.
Reported are measures of Subjective Wellbeing abd Body Satisfaction.
}
\keyword{datasets}
