% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bodywellm}
\alias{data_bodywellm}
\title{BodyWellM - Ch12 - Body Satisfaction and Well-being data for males from Figure 11.24 left panel}
\format{
\subsection{\code{data_bodywellm}}{

A data frame with 47 rows and 2 columns:
\describe{
\item{Body Satisfaction}{numeric}
\item{Well-being}{numeric}
}
}
}
\usage{
data_bodywellm
}
\description{
A subset of data_bodywell_fm, reports only for those participants
who identified as male.  Data is Subjective Wellbeing abd Body Satisfaction.
}
\keyword{datasets}
