% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_bushman_2005}
\alias{data_bushman_2005}
\title{Sample data for estimate_mdiff_ind_contrast: Bushman (2005)}
\format{
\subsection{\code{data_bushman_2005} A data frame with 3 rows and 4 columns:}{

\describe{
\item{condition}{Factor indicating group: Neutral, Violent, or Sexual}
\item{n}{Group sample sizes}
\item{m}{Group means}
\item{s}{Group standard deviation}
}
}
}
\source{
Bushman, B. (2005). Violence and sex in television programs do not
sell products in advertisements. \emph{Psychological Science}, 16, 702-–708.
doi:10.1111/j.1467-9280.2005.01599.x
}
\usage{
data_bushman_2005
}
\description{
Bushman (2005) reported an investigation of people’s memory for ads that were
presented during different types of television show. He wanted to estimate to
what extent violent content, compared with neutral content, might lead to
reduced memory for ads, and perhaps reduced purchasing intentions for
advertised products. He investigated the same questions for shows with sexual
content. He chose a three- group design as an efficient way to investigate
both types of content. A sample of 252 typical TV viewers were randomly
assigned to watch one of three types of show. Some watched a Neutral show
(e.g., America’s Funniest Animals), some a show with Violent content (e.g.,
Cops), and others a show with Sexual content (e.g., Sex in the City). The
viewers watched different shows, but all saw the same 12 ads inserted into
the shows. The ads were genuine advertisements, but for little-known
products, so most viewers had never seen the ads before. To enhance realism,
viewers watched the shows in easy chairs, with snacks and soda available.
After the viewing came a surprise memory test for the ads (participants had
not been told the purpose of the study). I’ll report data for memory
recognition: Participants saw a list of 12 products, with four 4 brand names
for each type of product, just one of which had appeared in an ad. For each
set of four 4 brands, participants had to choose which one they felt they
recognized from the ads they had just seen, so the maximum score was 12. This
is a summary data set, providing sample sizes, means, and standard deviations
for each group in the study.
}
\keyword{datasets}
