% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_campus_involvement}
\alias{data_campus_involvement}
\title{Campus Involvement - Ch11 - for End-of-Chapter Exercise 11.7}
\format{
\subsection{\code{data_campus_involvement}}{

A data frame with 113 rows and 6 columns:
\describe{
\item{ID}{integer}
\item{Gender}{factor}
\item{GPA}{numeric}
\item{CommuterStatus}{factor}
\item{SWB}{numeric}
\item{Campus Involvement}{numeric}
}
}
}
\usage{
data_campus_involvement
}
\description{
Clinton conducted a survey of college students to determine the extent to
which subjective well-being is related to campus involvement (Campus
Involvement data set on the book website). Participants completed a measure
of subjective well-being (scale from 1 to 5) and a measure of campus
involvement (scale from 1 to 5). Participants also reported gender (male or
female) and commuter status (resident or commuter).  Synthetic data simulated
to mimic survey data from a class project.
}
\keyword{datasets}
