% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_chap_8_paired_ex_8.18}
\alias{data_chap_8_paired_ex_8.18}
\title{\emph{Fictitious} data from an unrealistically small HEAT
study comparing scores for a single group of students before and after a
workshop on climate change.}
\format{
\subsection{\code{data_chap_8_paired_ex_8.18}}{

A data frame with 8 rows and 2 columns:
\describe{
\item{Before}{numeric}
\item{After}{numeric}
}
}
}
\usage{
data_chap_8_paired_ex_8.18
}
\description{
\emph{Fictitious} data from an unrealistically small HEAT
study comparing scores for a single group of students before and after a
workshop on climate change.
}
\keyword{datasets}
