% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_clean_moral}
\alias{data_clean_moral}
\title{Clean moral - Ch07 - from Schnall et al. (2008), Study 1, and Johnson et al. (2014)}
\format{
\subsection{\code{data_clean_moral}}{

A data frame with 208 rows and 4 columns:
\describe{
\item{Schnall Condition}{factor}
\item{Schnall Moral judgment}{numeric}
\item{Johnson Condition}{factor}
\item{Johnson Moral judgment}{numeric}
}
}
}
\usage{
data_clean_moral
}
\description{
Clean moral - Ch07 - from Schnall et al. (2008), Study 1, and Johnson et al. (2014)
}
\keyword{datasets}
