% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_college_survey_1}
\alias{data_college_survey_1}
\title{College survey 1 - Ch03 - for End-of-Chapter Exercise 3.3}
\format{
\subsection{\code{data_college_survey_1}}{

A data frame with 243 rows and 23 columns:
\describe{
\item{ID}{integer}
\item{Gender}{factor}
\item{Gender_Code}{factor}
\item{Age}{integer}
\item{Shool_Year}{factor}
\item{School_Year_Code}{factor}
\item{Transfer}{factor}
\item{Transfer_Code}{logical}
\item{Student_Athlete}{factor}
\item{Student_Athlete_Code}{logical}
\item{Wealth_SR}{numeric}
\item{GPA}{numeric}
\item{ACT}{integer}
\item{Subjective_Well_Being}{numeric}
\item{Positive_Affect}{numeric}
\item{Negative_Affect}{numeric}
\item{Relationship_Confidence}{numeric}
\item{Exercise}{numeric}
\item{Academic_Motivation_Intrinsic}{numeric}
\item{Academic_Motivation_Extrinsic}{numeric}
\item{Academic_Motivation_Amotivation}{numeric}
\item{Intelligence_Value}{numeric}
\item{Raven_Score}{numeric}
}
}
}
\usage{
data_college_survey_1
}
\description{
Data from an additional survey of Dominican University students; reports
various psychological and behavioral measures.
}
\keyword{datasets}
