% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_college_survey_2}
\alias{data_college_survey_2}
\title{College survey 2 - Ch05 - for End-of-Chapter Exercise 5.4}
\format{
\subsection{\code{data_college_survey_2}}{

A data frame with 138 rows and 17 columns:
\describe{
\item{ID}{integer}
\item{Gender}{factor}
\item{Gender_Code}{factor}
\item{Age}{numeric}
\item{Wealth_SR}{numeric}
\item{School_Year}{factor}
\item{School_Year_Code}{factor}
\item{Transfer}{factor}
\item{Transfer_Code}{logical}
\item{GPA}{numeric}
\item{Subjective_Well_Being}{numeric}
\item{Positive_Affect}{numeric}
\item{Negative_Affect}{numeric}
\item{Academic_Engagement}{numeric}
\item{Religious_Meaning}{numeric}
\item{Health}{numeric}
\item{Emotion_Recognition}{factor}
}
}
}
\usage{
data_college_survey_2
}
\description{
College survey 2 - Ch05 - for End-of-Chapter Exercise 5.4
}
\keyword{datasets}
