% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_damischrcj}
\alias{data_damischrcj}
\title{DamischRCJ - Ch9 - from 6 Damisch studies, and Calin-Jageman and Caldwell (2014)}
\format{
\subsection{\code{data_damischrcj}}{

A data frame with 8 rows and 5 columns:
\describe{
\item{Study}{factor}
\item{Cohen's d unbiased}{numeric}
\item{n Control}{integer}
\item{n Lucky}{integer}
\item{Research Group}{factor}
}
}
}
\usage{
data_damischrcj
}
\description{
DamischRCJ - Ch9 - from 6 Damisch studies, and Calin-Jageman and Caldwell (2014)
}
\keyword{datasets}
