% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_effronraj_fakenews}
\alias{data_effronraj_fakenews}
\title{EffronRaj fakenews - Ch8 - from Effron and Raj (2020)}
\format{
\subsection{\code{data_effronraj_fakenews}}{

A data frame with 138 rows and 5 columns:
\describe{
\item{ID}{factor}
\item{UnethOld}{numeric}
\item{UnethNew}{numeric}
\item{AccurOld}{numeric}
\item{AccurNew}{numeric}
}
}
}
\usage{
data_effronraj_fakenews
}
\description{
EffronRaj fakenews - Ch8 - from Effron and Raj (2020)
}
\keyword{datasets}
