% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_emotion_heartrate}
\alias{data_emotion_heartrate}
\title{Emotion heartrate - Ch8 - from Lakens (2013)}
\format{
\subsection{\code{data_emotion_heartrate}}{

A data frame with 68 rows and 3 columns:
\describe{
\item{ID}{integer}
\item{HR_baseline}{numeric}
\item{HR_anger}{numeric}
}
}
}
\usage{
data_emotion_heartrate
}
\description{
Emotion heartrate - Ch8 - from Lakens (2013)
}
\keyword{datasets}
