% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_exam_scores}
\alias{data_exam_scores}
\title{Exam Scores - Ch11 - for End-of-Chapter Exercise 11.2}
\format{
\subsection{\code{data_exam_scores}}{

A data frame with 9 rows and 3 columns:
\describe{
\item{StudentID}{factor}
\item{First Exam}{numeric}
\item{Final Exam}{numeric}
}
}
}
\usage{
data_exam_scores
}
\description{
To what extent does initial performance in a class relate to performance on a
final exam? First exam and final exam scores for nine students enrolled in an
introductory psychology course. Exam scores are percentages, where 0 = no
answers correct and 100 = all answers correct.
Data is synthetic to represent patterns found in a previous psych
stats course.
}
\keyword{datasets}
