% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_flag_priming_ma}
\alias{data_flag_priming_ma}
\title{Flag Priming ma - Ch9 - Many Labs replications of Carter et al. (2011)}
\format{
\subsection{\code{data_flag_priming_ma}}{

A data frame with 25 rows and 7 columns:
\describe{
\item{Location}{factor}
\item{M Flag}{numeric}
\item{s Flag}{numeric}
\item{n Flag}{integer}
\item{M Noflag}{numeric}
\item{s Noflag}{numeric}
\item{n Noflag}{integer}
}
}
}
\usage{
data_flag_priming_ma
}
\description{
Flag Priming ma - Ch9 - Many Labs replications of Carter et al. (2011)
}
\keyword{datasets}
