% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gender_math_iat}
\alias{data_gender_math_iat}
\title{Gender math IAT - Ch07 - Ithaca and SDSU replications of Nosek et al. (2002)}
\format{
\subsection{\code{data_gender_math_iat}}{

A data frame with 155 rows and 4 columns:
\describe{
\item{Ithaca gender}{factor}
\item{Ithaca IAT}{numeric}
\item{SDSU gender}{factor}
\item{SDSU IAT}{numeric}
}
}
}
\usage{
data_gender_math_iat
}
\description{
Gender math IAT - Ch07 - Ithaca and SDSU replications of Nosek et al. (2002)
}
\keyword{datasets}
