% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gender_math_iat_ma}
\alias{data_gender_math_iat_ma}
\title{Gender math IAT ma - Ch9 - Many Labs replications of Nosek et al. (2002)}
\format{
\subsection{\code{data_gender_math_iat_ma}}{

A data frame with 30 rows and 9 columns:
\describe{
\item{Location}{factor}
\item{M Male}{numeric}
\item{s Male}{numeric}
\item{n Male}{integer}
\item{M Female}{numeric}
\item{s Female}{numeric}
\item{n Female}{integer}
\item{USAorNot}{factor}
\item{Country}{factor}
}
}
}
\usage{
data_gender_math_iat_ma
}
\description{
Gender math IAT ma - Ch9 - Many Labs replications of Nosek et al. (2002)
}
\keyword{datasets}
