% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_halagappa}
\alias{data_halagappa}
\title{Halagappa - Ch14 - from Halagappa et al. (2007)}
\format{
\subsection{\code{data_halagappa}}{

A data frame with 6 rows and 4 columns:
\describe{
\item{Groups}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{n}{integer}
}
}
}
\usage{
data_halagappa
}
\description{
Halagappa - Ch14 - from Halagappa et al. (2007)
}
\keyword{datasets}
