% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_halagappa_et_al_2007}
\alias{data_halagappa_et_al_2007}
\title{Sample data for estimate_mdiff_ind_contrast: Halagappa (2007)}
\format{
\subsection{\code{data_halagappa_et_al_2007} A data frame with 3 rows and 4}{

columns:
\describe{
\item{condition}{Factor indicating group: NFree10, AFree10, ADiet10, NFree17, AFree17, ADiet17}
\item{n}{Group sample sizes}
\item{m}{Group means}
\item{s}{Group standard deviation}
}
}
}
\source{
Halagappa, V. K. M., Guo, Z., Pearson, M., Matsuoka, Y., Cutler, R.
G., LaFerla, F. M., & Mattson, M. P. (2007). Intermittent fasting and
caloric restriction ameliorate age-related behavioral deficits in the
triple-transgenic mouse model of Alzheimer's disease.
\emph{Neurobiology of Disease}, 26, 212–220. doi:10.1016/j.nbd.2006.12.019
}
\usage{
data_halagappa_et_al_2007
}
\description{
Could eating much less delay Alzheimer’s? If so, that would be great news.
Halagappa et al. (2007) investigated the possibility by using a mouse model,
meaning they used Alzheimer-prone mice, which were genetically predisposed to
develop neural degeneration typical of Alzheimer’s. The researchers used six
independent groups of mice, three tested in mouse middle age when 10 months
old, and three in mouse old age when 17 months. At each age there was a
control group of normal mice that ate freely (the NFree10 and NFree17
groups), a group of Alzheimer-prone mice that also ate freely (the AFree10
and AFree17 groups), and another Alzheimer-prone group restricted to 40\% less
food than normal (the ADiet10 and ADiet17 groups). Table 14.2 lists the
factors that define the groups, the group labels, and their means, M1, M2, …
(as we’ll see them displayed in ESCI). I’ll discuss one measure of mouse
cognition: the percent time spent near the target of a water maze, with
higher values indicating better learning and memory. This is a summary data
set, providing sample sizes, means, and standard deviations for each group in
the study.
}
\keyword{datasets}
