% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_home_prices}
\alias{data_home_prices}
\title{Home Prices - Ch12 - for End-of-Chapter Exercise 12.2}
\format{
\subsection{\code{data_home_prices}}{

A data frame with 300 rows and 8 columns:
\describe{
\item{MLS}{integer}
\item{Location}{factor}
\item{Price}{integer}
\item{Bedrooms}{integer}
\item{Bathrooms}{integer}
\item{Size (m2)}{numeric}
\item{Status}{factor}
\item{Status_Code}{logical}
}
}
}
\usage{
data_home_prices
}
\description{
Home Prices - Ch12 - for End-of-Chapter Exercise 12.2
}
\keyword{datasets}
