% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_kardas_expt_3}
\alias{data_kardas_expt_3}
\title{Kardas Expt 3 - Ch07 - from Kardas and O’Brien (2018), Experiment 3}
\format{
\subsection{\code{data_kardas_expt_3}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{Exposure}{factor}
\item{Prediction}{numeric}
\item{Performance}{numeric}
}
}
}
\usage{
data_kardas_expt_3
}
\description{
Kardas Expt 3 - Ch07 - from Kardas and O’Brien (2018), Experiment 3
}
\keyword{datasets}
