% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_kardas_expt_4}
\alias{data_kardas_expt_4}
\title{Kardas Expt 4 - Ch07 - from Kardas and O’Brien (2018), Experiment 4}
\format{
\subsection{\code{data_kardas_expt_4}}{

A data frame with 270 rows and 4 columns:
\describe{
\item{Exposure}{factor}
\item{Prediction}{integer}
\item{Performance}{integer}
\item{Time}{numeric}
}
}
}
\usage{
data_kardas_expt_4
}
\description{
Kardas Expt 4 - Ch07 - from Kardas and O’Brien (2018), Experiment 4
}
\keyword{datasets}
