% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_labels_flavor}
\alias{data_labels_flavor}
\title{Labels flavor - Ch8 - from Floretta-Schiller et al. (2015)}
\format{
\subsection{\code{data_labels_flavor}}{

A data frame with 51 rows and 6 columns:
\describe{
\item{ParticipantID}{integer}
\item{Enjoy_Generic}{numeric}
\item{Enjoy_Organic}{numeric}
\item{Pay_Generic}{numeric}
\item{Pay_Organic}{numeric}
\item{Suspicious}{logical}
}
}
}
\usage{
data_labels_flavor
}
\description{
Labels flavor - Ch8 - from Floretta-Schiller et al. (2015)
}
\keyword{datasets}
