% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_latimier_3groups}
\alias{data_latimier_3groups}
\title{Latimier 3Groups - Ch14 - 3 groups in Latimier et al. (2019)}
\format{
\subsection{\code{data_latimier_3groups}}{

A data frame with 285 rows and 3 columns:
\describe{
\item{Group}{factor}
\item{Test\%}{numeric}
\item{Time}{numeric}
}
}
}
\usage{
data_latimier_3groups
}
\description{
Latimier 3Groups - Ch14 - 3 groups in Latimier et al. (2019)
}
\keyword{datasets}
