% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_latimier_prequiz}
\alias{data_latimier_prequiz}
\title{Latimier Prequiz - Ch03 - Prequiz group in Latimier et al. (2019)}
\format{
\subsection{\code{data_latimier_prequiz}}{

A data frame with 95 rows and 3 columns:
\describe{
\item{Group}{factor}
\item{Test\%}{numeric}
\item{TIme}{numeric}
}
}
}
\usage{
data_latimier_prequiz
}
\description{
Latimier Prequiz - Ch03 - Prequiz group in Latimier et al. (2019)
}
\keyword{datasets}
