% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_macnamara_r_ma}
\alias{data_macnamara_r_ma}
\title{Macnamara r ma - Ch11 - from Macnamara et al. (2014)}
\format{
\subsection{\code{data_macnamara_r_ma}}{

A data frame with 16 rows and 4 columns:
\describe{
\item{Study}{factor}
\item{r}{numeric}
\item{N}{integer}
\item{Instrument_Type}{factor}
}
}
}
\usage{
data_macnamara_r_ma
}
\description{
Macnamara r ma - Ch11 - from Macnamara et al. (2014)
}
\keyword{datasets}
