% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mccabemichael_brain}
\alias{data_mccabemichael_brain}
\title{McCabeMichael brain - Ch9 - from Michael et al. (2013)}
\format{
\subsection{\code{data_mccabemichael_brain}}{

A data frame with 10 rows and 9 columns:
\describe{
\item{Study name}{factor}
\item{M No Brain}{numeric}
\item{s No Brain}{numeric}
\item{n No Brain}{numeric}
\item{M Brain}{numeric}
\item{s Brain}{numeric}
\item{n Brain}{numeric}
\item{SimpleCritique}{factor}
\item{Research group}{factor}
}
}
}
\usage{
data_mccabemichael_brain
}
\description{
McCabeMichael brain - Ch9 - from Michael et al. (2013)
}
\keyword{datasets}
