% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mccabemichael_brain2}
\alias{data_mccabemichael_brain2}
\title{McCabeMichael brain2 - Ch9 - from Michael et al. (2013)}
\format{
\subsection{\code{data_mccabemichael_brain2}}{

A data frame with 12 rows and 9 columns:
\describe{
\item{Study name}{factor}
\item{M No Brain}{numeric}
\item{s No Brain}{numeric}
\item{n No Brain}{numeric}
\item{M Brain}{numeric}
\item{s Brain}{numeric}
\item{n Brain}{numeric}
\item{SimpleCritique}{factor}
\item{Research group}{factor}
}
}
}
\source{
\doi{10.3758/s13423-013-0391-6}
}
\usage{
data_mccabemichael_brain2
}
\description{
Same as data_mccabemichael_brain but includes two additional critique studies
run by the Michael team.
}
\keyword{datasets}
