% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_meditationbrain}
\alias{data_meditationbrain}
\title{MeditationBrain - Ch15 - from Hölzel et al. (2011)}
\format{
\subsection{\code{data_meditationbrain}}{

A data frame with 33 rows and 7 columns:
\describe{
\item{Pretest}{numeric}
\item{Posttest}{numeric}
\item{Group}{factor}
\item{ControlPre}{numeric}
\item{ControlPost}{numeric}
\item{MeditationPre}{numeric}
\item{MeditationPost}{numeric}
}
}
}
\usage{
data_meditationbrain
}
\description{
MeditationBrain - Ch15 - from Hölzel et al. (2011)
}
\keyword{datasets}
