% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_organicmoral}
\alias{data_organicmoral}
\title{OrganicMoral - Ch14 - from Eskine (2013)}
\format{
\subsection{\code{data_organicmoral}}{

A data frame with 106 rows and 6 columns:
\describe{
\item{Group}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{N}{integer}
\item{ReplicationGroup}{factor}
\item{MoralJudgmentment}{numeric}
}
}
}
\usage{
data_organicmoral
}
\description{
OrganicMoral - Ch14 - from Eskine (2013)
}
\keyword{datasets}
