% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_penlaptop1}
\alias{data_penlaptop1}
\title{Sample data for estimate_mdiff_two: Transcription scores from Study 1 of
Mueller & Oppenheimer (2014)}
\format{
\subsection{\code{data_penlaptop1} A data frame with 65 rows and 2 columns:}{

\describe{
\item{condition}{A factor indicating membership in the pen or laptop group}
\item{transcription}{\% of notes transcribed verbatim from lecture}
}
}
}
\source{
\url{https://osf.io/23yad}

Mueller, P. A., & Oppenheimer, D. M. (2014). The Pen Is Mightier Than
the Keyboard: Advantages of Longhand Over Laptop Note Taking.
\emph{Psychological Science}, 25(6), 1159–1168.
\url{https://doi.org/10.1177/0956797614524581}
}
\usage{
data_penlaptop1
}
\description{
Mueller & Oppenheimer (2014) asked students to take notes on a video lecture
using a pen or a laptop.  One dependent measure was \% transcription
(verbatim) copying, measured as the number of 3-gram matches to the text of
the video lecture. This data set is the \% transcription for the pen and
laptop groups.  It is suitable for estimating differences between two
independent groups for a continuous measure.
}
\keyword{datasets}
