% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_powerperformance_ma}
\alias{data_powerperformance_ma}
\title{PowerPerformance ma - Ch9 - from Burgmer and Englich (2012), and Cusack et al. (2015)}
\format{
\subsection{\code{data_powerperformance_ma}}{

A data frame with 8 rows and 12 columns:
\describe{
\item{StudyName}{factor}
\item{Country}{factor}
\item{Population}{factor}
\item{Difficulty}{factor}
\item{Task}{factor}
\item{M Control}{numeric}
\item{s Control}{numeric}
\item{M Power}{numeric}
\item{s Power}{numeric}
\item{Cohen d unb}{numeric}
\item{n Control}{integer}
\item{n Power}{integer}
}
}
}
\usage{
data_powerperformance_ma
}
\description{
PowerPerformance ma - Ch9 - from Burgmer and Englich (2012), and Cusack et al. (2015)
}
\keyword{datasets}
