% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_rattanmotivation}
\alias{data_rattanmotivation}
\title{RattanMotivation - Ch14 - from Rattan et al. (2012)}
\format{
\subsection{\code{data_rattanmotivation}}{

A data frame with 54 rows and 2 columns:
\describe{
\item{Group}{factor}
\item{Motivation}{numeric}
}
}
}
\usage{
data_rattanmotivation
}
\description{
RattanMotivation - Ch14 - from Rattan et al. (2012)
}
\keyword{datasets}
