% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_religionsharing}
\alias{data_religionsharing}
\title{ReligionSharing - Ch14 - for End-of-Chapter Exercise 14.3}
\format{
\subsection{\code{data_religionsharing}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{Group}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{N}{integer}
}
}
}
\usage{
data_religionsharing
}
\description{
ReligionSharing - Ch14 - for End-of-Chapter Exercise 14.3
}
\keyword{datasets}
