% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_religious_belief}
\alias{data_religious_belief}
\title{Religious belief - Ch03 - for End-of-Chapter Exercise 3.5}
\format{
\subsection{\code{data_religious_belief}}{

A data frame with 213 rows and 3 columns:
\describe{
\item{Response_ID}{character}
\item{Belief_in_God}{factor}
\item{Age}{integer}
}
}
}
\usage{
data_religious_belief
}
\description{
Let's look at some data about religious beliefs. The Religious belief file
has data from a large online survey in which participants were asked to
report, on a scale from 0 to 100, their belief in the existence of God.  Age
was also reported.
}
\keyword{datasets}
