% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_selfexplain}
\alias{data_selfexplain}
\title{SelfExplain - Ch15 - from McEldoon et al. (2013)}
\format{
\subsection{\code{data_selfexplain}}{

A data frame with 52 rows and 4 columns:
\describe{
\item{Student ID}{factor}
\item{Condition}{factor}
\item{Pretest}{numeric}
\item{Posttest}{numeric}
}
}
}
\usage{
data_selfexplain
}
\description{
SelfExplain - Ch15 - from McEldoon et al. (2013)
}
\keyword{datasets}
