% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_simmonscredibility}
\alias{data_simmonscredibility}
\title{SimmonsCredibility - Ch14 - from Simmons and Nelson (2020)}
\format{
\subsection{\code{data_simmonscredibility}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{Groups}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{n}{integer}
}
}
}
\usage{
data_simmonscredibility
}
\description{
SimmonsCredibility - Ch14 - from Simmons and Nelson (2020)
}
\keyword{datasets}
