% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sleep_beauty}
\alias{data_sleep_beauty}
\title{Sleep Beauty - Ch11 - for End-of-Chapter Exercise 11.6}
\format{
\subsection{\code{data_sleep_beauty}}{

A data frame with 70 rows and 2 columns:
\describe{
\item{Sleep (hours)}{numeric}
\item{Rated_Attractiveness}{numeric}
}
}
}
\usage{
data_sleep_beauty
}
\description{
Sleep Beauty - Ch11 - for End-of-Chapter Exercise 11.6
}
\keyword{datasets}
