% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_smithrecall}
\alias{data_smithrecall}
\title{SmithRecall - Ch15 - from Smith et al. (2016)}
\format{
\subsection{\code{data_smithrecall}}{

A data frame with 120 rows and 6 columns:
\describe{
\item{ID}{integer}
\item{Study Technique}{factor}
\item{Stress Status}{factor}
\item{\%Recalled}{numeric}
\item{Items Recalled}{numeric}
\item{Group}{factor}
}
}
}
\usage{
data_smithrecall
}
\description{
SmithRecall - Ch15 - from Smith et al. (2016)
}
\keyword{datasets}
