% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_stickgold}
\alias{data_stickgold}
\title{Stickgold - Ch06 - from Stickgold et al. (2000)}
\format{
\subsection{\code{data_stickgold}}{

A data frame with 11 rows and 3 columns:
\describe{
\item{Sleep deprived}{numeric}
\item{B}{factor}
\item{C}{factor}
}
}
}
\usage{
data_stickgold
}
\description{
Stickgold - Ch06 - from Stickgold et al. (2000)
}
\keyword{datasets}
