% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_studystrategies}
\alias{data_studystrategies}
\title{StudyStrategies - Ch14 - from O’Reilly et al. (1998)}
\format{
\subsection{\code{data_studystrategies}}{

A data frame with 3 rows and 10 columns:
\describe{
\item{Group}{factor}
\item{TestMean}{numeric}
\item{TestSD}{numeric}
\item{TestN}{integer}
\item{PrevKnowMean}{numeric}
\item{PrevKnowSD}{numeric}
\item{PrevKnowN}{integer}
\item{EaseUseMean}{numeric}
\item{EaseUseSD}{numeric}
\item{EaseUseN}{integer}
}
}
}
\usage{
data_studystrategies
}
\description{
StudyStrategies - Ch14 - from O’Reilly et al. (1998)
}
\keyword{datasets}
