% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_thomason_1}
\alias{data_thomason_1}
\title{Thomason 1 - Ch11 - from Thomason 1}
\format{
\subsection{\code{data_thomason_1}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{Participant ID}{factor}
\item{Pretest}{numeric}
\item{Posttest}{numeric}
}
}
}
\usage{
data_thomason_1
}
\description{
Thomason 1 - Ch11 - from Thomason 1
}
\keyword{datasets}
