% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_videogameaggression}
\alias{data_videogameaggression}
\title{VideogameAggression - Ch15 - from Hilgard (2015)}
\format{
\subsection{\code{data_videogameaggression}}{

A data frame with 223 rows and 3 columns:
\describe{
\item{Violence}{factor}
\item{Difficulty}{factor}
\item{Agression}{numeric}
}
}
}
\usage{
data_videogameaggression
}
\description{
VideogameAggression - Ch15 - from Hilgard (2015)
}
\keyword{datasets}
