% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_magnitude.R
\name{estimate_magnitude}
\alias{estimate_magnitude}
\title{Estimate a magnitude.}
\usage{
estimate_magnitude(
  data = NULL,
  outcome_variable = NULL,
  mean = NULL,
  sd = NULL,
  n = NULL,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{mean}{For summary data - A numeric}

\item{sd}{For summary data - A numeric > 0}

\item{n}{For summary data - An integer > 0}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returnsobject of class esci_estimate
}
\description{
Returns effect sizes appropriate for estimating properties from a
quantitative variable.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_magnitude(
 datasets::PlantGrowth,
 weight
)

}
