% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pdiff_one.R
\name{estimate_pdiff_one}
\alias{estimate_pdiff_one}
\title{Estimate magnitudes of difference from a single measure compared to a know reference.}
\usage{
estimate_pdiff_one(
  data = NULL,
  outcome_variable = NULL,
  comparison_cases = NULL,
  comparison_n = NULL,
  reference_p = 0,
  case_label = 1,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  count_NA = FALSE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, which must be a factor, or a vector that is a factor}

\item{comparison_cases}{For summary data, a numeric integer > 0}

\item{comparison_n}{For summary data, a numeric integer >= count}

\item{reference_p}{Reference proportion, numeric >=0 and <=1}

\item{case_label}{An optional numeric or character label for the
count level.}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{count_NA}{Logical to count NAs (TRUE) in total N or not (FALSE)}
}
\value{
Returns object of class esci_estimate
}
\description{
Returns effect sizes appropriate for designs with one quantitative variable
compared against a known reference.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_pdiff_one(
 datasets::PlantGrowth,
 group,
 reference_p = 0.33
)

}
