% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_proportion.R
\name{estimate_proportion}
\alias{estimate_proportion}
\title{Estimate a proportion.}
\usage{
estimate_proportion(
  data = NULL,
  outcome_variable = NULL,
  cases = NULL,
  case_label = 1,
  outcome_variable_levels = NULL,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  count_NA = FALSE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, which must be a factor, or a vector that is a factor}

\item{cases}{For summary data - A vector of cases}

\item{case_label}{A numeric or string indicating which level
of the factor to estimate.  Defaults to 1, meaning first level is analyzed}

\item{outcome_variable_levels}{For summary data - optional vector of 2
characters indicating name of the count level and name of the not count
level.  Defaults to "Affected" and "Not Affected"}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{count_NA}{Logical to count NAs (TRUE) in total N or not (FALSE)}
}
\value{
Returnsobject of class esci_estimate
}
\description{
Returns effect sizes appropriate for estimating properties from a
nominal variable.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_proportion(
 datasets::PlantGrowth,
 group
)

}
