% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTwoProportions.R
\name{estimateProportion}
\alias{estimateProportion}
\title{Estimate a proportion}
\usage{
estimateProportion(cases, n, caselabels = c("Affected", "Not Affected"),
  conf.level = 0.95)
}
\arguments{
\item{cases}{The number of observations of the specific outcome}

\item{n}{The total number of observations}

\item{caselabels}{Optional argument to define labels for outcomes.  Default is "Affected" and "Not Affected".}

\item{conf.level}{Confidence level for the confidence interval.  Default is 0.95.  Must be >0.5 and <1.}
}
\description{
\code{estimateProportion} estimates a proportion along with a confidence interval for a categorical variable.
The function can accept a column of categorical outcomes (a factor).  It will calculate the proportion of 
outcomes of a specified level (cases) out of all valid outcomes (n): P = cases/N.  It will then calculate
the confidence interval on that proportion.
}
\details{
This function can also accept summary data, where the user pases the number of cases and sample size directly.
}
\examples{
estimateProportion(cases = 10, n = 100)
estimateProportion(cases = 10, n = 100, conf.level = 0.90)
estimateProportion(cases = 10, n = 100, caselabels = c("Employed", "Unemployed"))
}
