% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovidescribe.h.R
\name{jamovidescribe}
\alias{jamovidescribe}
\title{Describe}
\usage{
jamovidescribe(
  data,
  outcome_variable,
  show_details = FALSE,
  mark_mean = FALSE,
  mark_median = FALSE,
  mark_sd = FALSE,
  mark_quartiles = FALSE,
  mark_z_lines = FALSE,
  mark_percentile = "0",
  histogram_bins = "12",
  es_plot_width = "500",
  es_plot_height = "400",
  ymin = "auto",
  ymax = "auto",
  breaks = "auto",
  xmin = "auto",
  xmax = "auto",
  xbreaks = "auto",
  ylab = "auto",
  xlab = "auto",
  axis.text.y = "14",
  axis.title.y = "15",
  axis.text.x = "14",
  axis.title.x = "15",
  fill_regular = "#008DF9",
  fill_highlighted = "#E20134",
  color = "black"
)
}
\arguments{
\item{data}{.}

\item{outcome_variable}{.}

\item{show_details}{.}

\item{mark_mean}{.}

\item{mark_median}{.}

\item{mark_sd}{.}

\item{mark_quartiles}{.}

\item{mark_z_lines}{.}

\item{mark_percentile}{.}

\item{histogram_bins}{.}

\item{es_plot_width}{.}

\item{es_plot_height}{.}

\item{ymin}{.}

\item{ymax}{.}

\item{breaks}{.}

\item{xmin}{.}

\item{xmax}{.}

\item{xbreaks}{.}

\item{ylab}{.}

\item{xlab}{.}

\item{axis.text.y}{.}

\item{axis.title.y}{.}

\item{axis.text.x}{.}

\item{axis.title.x}{.}

\item{fill_regular}{.}

\item{fill_highlighted}{.}

\item{color}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$describe_plot_warnings} \tab \tab \tab \tab \tab a html \cr
\code{results$describe_plot} \tab \tab \tab \tab \tab an image \cr
\code{results$describe_dotplot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Describe
}
