% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimdiff2x2.h.R
\name{jamovimdiff2x2}
\alias{jamovimdiff2x2}
\title{Means and Medians: 2x2 Factorial}
\usage{
jamovimdiff2x2(
  design = "fully_between",
  switch = "from_raw",
  data,
  outcome_variable,
  grouping_variable_A,
  grouping_variable_B,
  outcome_variable_level1,
  outcome_variable_level2,
  outcome_variable_name_bs = "My outcome variable",
  grouping_variable,
  repeated_measures_name = "Time",
  outcome_variable_name = "My outcome variable",
  A1_label = "A1 level",
  A2_label = "A2 level",
  B1_label = "B1 level",
  B2_label = "B2 level",
  A_label = "Variable A",
  B_label = "Variable B",
  A1B1_mean = " ",
  A1B1_sd = " ",
  A1B1_n = " ",
  A1B2_mean = " ",
  A1B2_sd = " ",
  A1B2_n = " ",
  A2B1_mean = " ",
  A2B1_sd = " ",
  A2B1_n = " ",
  A2B2_mean = " ",
  A2B2_sd = " ",
  A2B2_n = " ",
  conf_level = 95,
  effect_size = "mean_difference",
  assume_equal_variance = TRUE,
  show_details = FALSE,
  show_interaction_plot = FALSE,
  show_CI = FALSE,
  evaluate_hypotheses = FALSE,
  null_value = "0",
  null_boundary = "0",
  rope_units = "raw",
  alpha = 0.05,
  null_color = "#A40122",
  es_plot_width = "700",
  es_plot_height = "400",
  ymin = "auto",
  ymax = "auto",
  ybreaks = "auto",
  ylab = "auto",
  xlab = "auto",
  axis.text.y = "14",
  axis.title.y = "15",
  axis.text.x = "14",
  axis.title.x = "15",
  simple_contrast_labels = FALSE,
  error_layout = "halfeye",
  error_scale = "0.25",
  error_nudge = "0.5",
  data_layout = "random",
  data_spread = "0.20",
  difference_axis_units = "raw",
  difference_axis_breaks = "auto",
  shape_raw_reference = "circle filled",
  shape_raw_comparison = "circle filled",
  shape_raw_unused = "circle filled",
  shape_summary_reference = "circle filled",
  shape_summary_comparison = "circle filled",
  shape_summary_unused = "circle filled",
  shape_summary_difference = "triangle filled",
  color_raw_reference = "#008DF9",
  color_raw_comparison = "#009F81",
  color_raw_unused = "gray65",
  color_summary_reference = "#008DF9",
  color_summary_comparison = "#009F81",
  color_summary_unused = "gray65",
  color_summary_difference = "black",
  fill_raw_reference = "NA",
  fill_raw_comparison = "NA",
  fill_raw_unused = "NA",
  fill_summary_reference = "#008DF9",
  fill_summary_comparison = "#009F81",
  fill_summary_unused = "gray65",
  fill_summary_difference = "black",
  size_raw_reference = "2",
  size_raw_comparison = "2",
  size_raw_unused = "1",
  size_summary_reference = "4",
  size_summary_comparison = "4",
  size_summary_unused = "4",
  size_summary_difference = "4",
  alpha_raw_reference = "1",
  alpha_raw_comparison = "1",
  alpha_raw_unused = "1",
  alpha_summary_reference = "1",
  alpha_summary_comparison = "1",
  alpha_summary_unused = "1",
  alpha_summary_difference = "1",
  linetype_summary_reference = "solid",
  linetype_summary_comparison = "solid",
  linetype_summary_unused = "solid",
  linetype_summary_difference = "solid",
  color_interval_reference = "black",
  color_interval_comparison = "black",
  color_interval_unused = "gray65",
  color_interval_difference = "black",
  size_interval_reference = "3",
  size_interval_comparison = "3",
  size_interval_unused = "3",
  size_interval_difference = "3",
  alpha_interval_reference = "1",
  alpha_interval_comparison = "1",
  alpha_interval_unused = "1",
  alpha_interval_difference = "1",
  alpha_error_reference = "1",
  alpha_error_comparison = "1",
  alpha_error_unused = "1",
  alpha_error_difference = "1",
  fill_error_reference = "gray75",
  fill_error_comparison = "gray75",
  fill_error_unused = "gray75",
  fill_error_difference = "gray75"
)
}
\arguments{
\item{design}{.}

\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{grouping_variable_A}{.}

\item{grouping_variable_B}{.}

\item{outcome_variable_level1}{.}

\item{outcome_variable_level2}{.}

\item{outcome_variable_name_bs}{.}

\item{grouping_variable}{.}

\item{repeated_measures_name}{.}

\item{outcome_variable_name}{.}

\item{A1_label}{.}

\item{A2_label}{.}

\item{B1_label}{.}

\item{B2_label}{.}

\item{A_label}{.}

\item{B_label}{.}

\item{A1B1_mean}{.}

\item{A1B1_sd}{.}

\item{A1B1_n}{.}

\item{A1B2_mean}{.}

\item{A1B2_sd}{.}

\item{A1B2_n}{.}

\item{A2B1_mean}{.}

\item{A2B1_sd}{.}

\item{A2B1_n}{.}

\item{A2B2_mean}{.}

\item{A2B2_sd}{.}

\item{A2B2_n}{.}

\item{conf_level}{.}

\item{effect_size}{.}

\item{assume_equal_variance}{.}

\item{show_details}{.}

\item{show_interaction_plot}{.}

\item{show_CI}{.}

\item{evaluate_hypotheses}{.}

\item{null_value}{.}

\item{null_boundary}{.}

\item{rope_units}{.}

\item{alpha}{.}

\item{null_color}{.}

\item{es_plot_width}{.}

\item{es_plot_height}{.}

\item{ymin}{.}

\item{ymax}{.}

\item{ybreaks}{.}

\item{ylab}{.}

\item{xlab}{.}

\item{axis.text.y}{.}

\item{axis.title.y}{.}

\item{axis.text.x}{.}

\item{axis.title.x}{.}

\item{simple_contrast_labels}{.}

\item{error_layout}{.}

\item{error_scale}{.}

\item{error_nudge}{.}

\item{data_layout}{.}

\item{data_spread}{.}

\item{difference_axis_units}{.}

\item{difference_axis_breaks}{.}

\item{shape_raw_reference}{.}

\item{shape_raw_comparison}{.}

\item{shape_raw_unused}{.}

\item{shape_summary_reference}{.}

\item{shape_summary_comparison}{.}

\item{shape_summary_unused}{.}

\item{shape_summary_difference}{.}

\item{color_raw_reference}{.}

\item{color_raw_comparison}{.}

\item{color_raw_unused}{.}

\item{color_summary_reference}{.}

\item{color_summary_comparison}{.}

\item{color_summary_unused}{.}

\item{color_summary_difference}{.}

\item{fill_raw_reference}{.}

\item{fill_raw_comparison}{.}

\item{fill_raw_unused}{.}

\item{fill_summary_reference}{.}

\item{fill_summary_comparison}{.}

\item{fill_summary_unused}{.}

\item{fill_summary_difference}{.}

\item{size_raw_reference}{.}

\item{size_raw_comparison}{.}

\item{size_raw_unused}{.}

\item{size_summary_reference}{.}

\item{size_summary_comparison}{.}

\item{size_summary_unused}{.}

\item{size_summary_difference}{.}

\item{alpha_raw_reference}{.}

\item{alpha_raw_comparison}{.}

\item{alpha_raw_unused}{.}

\item{alpha_summary_reference}{.}

\item{alpha_summary_comparison}{.}

\item{alpha_summary_unused}{.}

\item{alpha_summary_difference}{.}

\item{linetype_summary_reference}{.}

\item{linetype_summary_comparison}{.}

\item{linetype_summary_unused}{.}

\item{linetype_summary_difference}{.}

\item{color_interval_reference}{.}

\item{color_interval_comparison}{.}

\item{color_interval_unused}{.}

\item{color_interval_difference}{.}

\item{size_interval_reference}{.}

\item{size_interval_comparison}{.}

\item{size_interval_unused}{.}

\item{size_interval_difference}{.}

\item{alpha_interval_reference}{.}

\item{alpha_interval_comparison}{.}

\item{alpha_interval_unused}{.}

\item{alpha_interval_difference}{.}

\item{alpha_error_reference}{.}

\item{alpha_error_comparison}{.}

\item{alpha_error_unused}{.}

\item{alpha_error_difference}{.}

\item{fill_error_reference}{.}

\item{fill_error_comparison}{.}

\item{fill_error_unused}{.}

\item{fill_error_difference}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$analysis_type} \tab \tab \tab \tab \tab a html \cr
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$es_median_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_mean_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$es_smd} \tab \tab \tab \tab \tab a table \cr
\code{results$point_null} \tab \tab \tab \tab \tab a table \cr
\code{results$interval_null} \tab \tab \tab \tab \tab a table \cr
\code{results$estimation_plot_warnings} \tab \tab \tab \tab \tab a html \cr
\code{results$main_effect_A} \tab \tab \tab \tab \tab an image \cr
\code{results$main_effect_B} \tab \tab \tab \tab \tab an image \cr
\code{results$interaction} \tab \tab \tab \tab \tab an image \cr
\code{results$interaction_plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Means and Medians: 2x2 Factorial
}
