% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovimetapdiff.h.R
\name{jamovimetapdiff}
\alias{jamovimetapdiff}
\title{Meta-Analysis: Difference in Proportions}
\usage{
jamovimetapdiff(
  data,
  reference_cases,
  reference_ns,
  comparison_cases,
  comparison_ns,
  labels,
  moderator,
  effect_label = "My effect",
  reported_effect_size = "RD",
  conf_level = 95,
  random_effects = "random_effects",
  include_PIs = FALSE,
  show_details = FALSE,
  es_plot_width = "600",
  es_plot_height = "750",
  size_base = "2",
  size_multiplier = "3",
  axis.text.y = "14",
  report_CIs = FALSE,
  meta_diamond_height = ".25",
  xlab = "auto",
  xmin = "auto",
  xmax = "auto",
  xbreaks = "auto",
  mark_zero = FALSE,
  axis.text.x = "14",
  axis.title.x = "15",
  dlab = "auto",
  dmin = "auto",
  dmax = "auto",
  dbreaks = "auto",
  shape_raw_reference = "square filled",
  shape_raw_comparison = "square filled",
  shape_summary_difference = "triangle filled",
  shape_raw_unused = "square filled",
  color_raw_reference = "#008DF9",
  color_raw_comparison = "#009F81",
  color_raw_unused = "gray65",
  color_summary_reference = "#008DF9",
  color_summary_comparison = "#009F81",
  color_summary_unused = "gray75",
  color_summary_difference = "black",
  color_summary_overall = "black",
  fill_raw_reference = "#008DF9",
  fill_raw_comparison = "#009F81",
  fill_raw_unused = "gray65",
  fill_summary_reference = "#008DF9",
  fill_summary_comparison = "#009F81",
  fill_summary_unused = "gray75",
  fill_summary_difference = "black",
  fill_summary_overall = "black",
  alpha_raw_reference = "1",
  alpha_raw_comparison = "1",
  alpha_raw_unused = "1",
  alpha_summary_reference = "1",
  alpha_summary_comparison = "1",
  alpha_summary_unused = "1",
  alpha_summary_difference = "1",
  alpha_summary_overall = "1",
  linetype_raw_reference = "solid",
  linetype_raw_comparison = "solid",
  linetype_summary_difference = "solid",
  linetype_raw_unused = "solid",
  color_interval_reference = "black",
  color_interval_comparison = "black",
  color_interval_difference = "black",
  color_interval_unused = "black",
  size_interval_reference = ".50",
  size_interval_comparison = ".50",
  size_interval_difference = ".50",
  size_interval_unused = ".50",
  alpha_interval_reference = "1",
  alpha_interval_comparison = "1",
  alpha_interval_difference = "1",
  alpha_interval_unused = "1"
)
}
\arguments{
\item{data}{.}

\item{reference_cases}{.}

\item{reference_ns}{.}

\item{comparison_cases}{.}

\item{comparison_ns}{.}

\item{labels}{.}

\item{moderator}{.}

\item{effect_label}{.}

\item{reported_effect_size}{.}

\item{conf_level}{.}

\item{random_effects}{.}

\item{include_PIs}{.}

\item{show_details}{.}

\item{es_plot_width}{.}

\item{es_plot_height}{.}

\item{size_base}{.}

\item{size_multiplier}{.}

\item{axis.text.y}{.}

\item{report_CIs}{.}

\item{meta_diamond_height}{.}

\item{xlab}{.}

\item{xmin}{.}

\item{xmax}{.}

\item{xbreaks}{.}

\item{mark_zero}{.}

\item{axis.text.x}{.}

\item{axis.title.x}{.}

\item{dlab}{.}

\item{dmin}{.}

\item{dmax}{.}

\item{dbreaks}{.}

\item{shape_raw_reference}{.}

\item{shape_raw_comparison}{.}

\item{shape_summary_difference}{.}

\item{shape_raw_unused}{.}

\item{color_raw_reference}{.}

\item{color_raw_comparison}{.}

\item{color_raw_unused}{.}

\item{color_summary_reference}{.}

\item{color_summary_comparison}{.}

\item{color_summary_unused}{.}

\item{color_summary_difference}{.}

\item{color_summary_overall}{.}

\item{fill_raw_reference}{.}

\item{fill_raw_comparison}{.}

\item{fill_raw_unused}{.}

\item{fill_summary_reference}{.}

\item{fill_summary_comparison}{.}

\item{fill_summary_unused}{.}

\item{fill_summary_difference}{.}

\item{fill_summary_overall}{.}

\item{alpha_raw_reference}{.}

\item{alpha_raw_comparison}{.}

\item{alpha_raw_unused}{.}

\item{alpha_summary_reference}{.}

\item{alpha_summary_comparison}{.}

\item{alpha_summary_unused}{.}

\item{alpha_summary_difference}{.}

\item{alpha_summary_overall}{.}

\item{linetype_raw_reference}{.}

\item{linetype_raw_comparison}{.}

\item{linetype_summary_difference}{.}

\item{linetype_raw_unused}{.}

\item{color_interval_reference}{.}

\item{color_interval_comparison}{.}

\item{color_interval_difference}{.}

\item{color_interval_unused}{.}

\item{size_interval_reference}{.}

\item{size_interval_comparison}{.}

\item{size_interval_difference}{.}

\item{size_interval_unused}{.}

\item{alpha_interval_reference}{.}

\item{alpha_interval_comparison}{.}

\item{alpha_interval_difference}{.}

\item{alpha_interval_unused}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a html \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$raw_data} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta} \tab \tab \tab \tab \tab a table \cr
\code{results$es_heterogeneity} \tab \tab \tab \tab \tab a table \cr
\code{results$es_meta_difference} \tab \tab \tab \tab \tab a table \cr
\code{results$estimation_plot_warnings} \tab \tab \tab \tab \tab a html \cr
\code{results$estimation_plots} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raw_data$asDF}

\code{as.data.frame(results$raw_data)}
}
\description{
Meta-Analysis: Difference in Proportions
}
