% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamoviproportion.h.R
\name{jamoviproportion}
\alias{jamoviproportion}
\title{Proportions: Single Group}
\usage{
jamoviproportion(
  switch = "from_raw",
  data,
  outcome_variable,
  cases = " ",
  not_cases = " ",
  case_label = "Affected",
  not_case_label = "Not Affected",
  outcome_variable_name = "My outcome variable",
  count_NA = FALSE,
  conf_level = 95,
  show_details = FALSE,
  plot_possible = FALSE,
  evaluate_hypotheses = FALSE,
  null_value = "0",
  null_boundary = "0",
  alpha = 0.05,
  null_color = "#A40122",
  es_plot_width = "300",
  es_plot_height = "400",
  ymin = "0",
  ymax = "1",
  breaks = "auto",
  ylab = "auto",
  xlab = "auto",
  axis.text.y = "14",
  axis.title.y = "15",
  axis.text.x = "14",
  axis.title.x = "15",
  error_layout = "none",
  shape_summary = "circle filled",
  color_summary = "#008DF9",
  fill_summary = "#008DF9",
  size_summary = "4",
  alpha_summary = "1",
  linetype_summary = "solid"
)
}
\arguments{
\item{switch}{.}

\item{data}{.}

\item{outcome_variable}{.}

\item{cases}{.}

\item{not_cases}{.}

\item{case_label}{.}

\item{not_case_label}{.}

\item{outcome_variable_name}{.}

\item{count_NA}{.}

\item{conf_level}{.}

\item{show_details}{.}

\item{plot_possible}{.}

\item{evaluate_hypotheses}{.}

\item{null_value}{.}

\item{null_boundary}{.}

\item{alpha}{.}

\item{null_color}{.}

\item{es_plot_width}{.}

\item{es_plot_height}{.}

\item{ymin}{.}

\item{ymax}{.}

\item{breaks}{.}

\item{ylab}{.}

\item{xlab}{.}

\item{axis.text.y}{.}

\item{axis.title.y}{.}

\item{axis.text.x}{.}

\item{axis.title.x}{.}

\item{error_layout}{.}

\item{shape_summary}{.}

\item{color_summary}{.}

\item{fill_summary}{.}

\item{size_summary}{.}

\item{alpha_summary}{.}

\item{linetype_summary}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$debug} \tab \tab \tab \tab \tab a preformatted \cr
\code{results$help} \tab \tab \tab \tab \tab a html \cr
\code{results$overview} \tab \tab \tab \tab \tab a table \cr
\code{results$point_null} \tab \tab \tab \tab \tab a table \cr
\code{results$interval_null} \tab \tab \tab \tab \tab a table \cr
\code{results$magnitude_plot_warnings} \tab \tab \tab \tab \tab a html \cr
\code{results$estimation_plots} \tab \tab \tab \tab \tab an array of images \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$overview$asDF}

\code{as.data.frame(results$overview)}
}
\description{
Proportions: Single Group
}
