% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_documentation.R
\name{jesci_document_data}
\alias{jesci_document_data}
\title{also promote kind, altruistic behavior. Brethel-Haurwitz and Marsh (2014)
examined this idea by collecting data on U.S. states. A Gallup poll in 2010
was used to measure each state's well-being index, a measure of mean
happiness for the state's residents on a scale from 0 to 100. Next, a kidney
donation database for 1999-2010 was used to figure out each state's rate
(number of donations per 1 million people) of non-directed kidney
donations-that's giving one kidney to a stranger, an extremely generous and
altruistic thing to do!
their classic study, Jacowitz and Kahneman (1995) asked participants to
estimate how many babies are born each day in the United States. Participants
were given either a low anchor (more than 100 babies/day) or a high anchor
(less than 50,000 babies/day). Those who saw the low anchor estimated many
fewer births/day than those who saw the high anchor, which suggests that the
wording can have a profound influence. The correct answer, as it happens, is
~11,000 births/day in 2014. To investigate the extent that these results are
replicable, the Many Labs project repeated this classic study at many
different labs around the world. You can find the summary data for 30 of
these labs in the Anchor Estimate ma data file
numerous other issues can be highly damaging, but are thriving in this social
media age. Trying to debunk a conspiracy theory by presenting facts and
evidence often doesn't work, alas. Psychological inoculation, also similar to
prebunking, presents a mild form of misinformation, preferably with
explanation, in the hope of building resistance to real-life fake news-a sort
of vaccine for fake news. The Bad News game is a spin-off from research on
psychological inoculation. Basol et al. (2020) assessed the possible
effectiveness of this game as a fake news vaccine. At getbadnews.com you can
click 'About' for information, or just start playing the game- it's easy and
maybe even fun. You encounter mock Twitter (now X) fake news messages that
illustrate common strategies for making fake news memorable or believable.
You make choices between messages and decide which ones to 'forward' as you
try to spread fake news while building your credibility score and number of
'followers'-rather like real life for a conspiracy theorist wanting to spread
the word. Compete with your friends for credibility and number of followers.
Basol's online participants first saw 18 fictitious fake news tweets and
rated each for reliability (accuracy, believability), and also rated their
confidence in that reliability rating. Both ratings were on a 1 to 7 scale.
Those in the BadNews group then played the game for 15 minutes, whereas those
in the Control group played Tetris. Then all once again gave reliability and
confidence ratings for the 18 tweets.
decade earlier, had been one of several outside experts invited to scrutinize
the laboratory and experimental procedures of parapsychology researcher
Charles Honorton. Bem not only judged them adequate, but joined the research
effort and became a coauthor with Honorton. Bem and Honorton (1994) first
reviewed early ganzfeld studies and described how the experimental procedure
had been improved to reduce the chance that results could be influenced by
various possible biases, or leakages of information from sender to receiver.
For example, the randomization procedure was carried out automatically by
computer, and all stimuli were presented under computer control. Bem and
Honorton then presented data from studies conducted with the improved
procedure. Table 13.1 presents basic data from 10 studies reported by Bem and
Honorton (1994). Participants each made a single judgment, so in Pilot 1, for
example, 22 participants responded, with 8 of them giving a correct response.
Three pilot studies helped refine the procedures, then four studies used
novice receivers. Study 5 used 20 students of music, drama, or dance as
receivers, in response to suggestions that creative people might be more
likely to show telepathy. Studies 6 and 7 used receivers who had participated
in an earlier study. The proportion of hits expected by chance is .25, and
Table 13.1 shows that all but Study 1 found proportions higher than .25.
who identified as female.  Data is Subjective Wellbeing abd Body Satisfaction.
who identified as male.  Data is Subjective Wellbeing abd Body Satisfaction.
Reported are measures of Subjective Wellbeing abd Body Satisfaction.}
\usage{
jesci_document_data(save_files = FALSE)
}
\description{
also promote kind, altruistic behavior. Brethel-Haurwitz and Marsh (2014)
examined this idea by collecting data on U.S. states. A Gallup poll in 2010
was used to measure each state's well-being index, a measure of mean
happiness for the state's residents on a scale from 0 to 100. Next, a kidney
donation database for 1999-2010 was used to figure out each state's rate
(number of donations per 1 million people) of non-directed kidney
donations-that's giving one kidney to a stranger, an extremely generous and
altruistic thing to do!
their classic study, Jacowitz and Kahneman (1995) asked participants to
estimate how many babies are born each day in the United States. Participants
were given either a low anchor (more than 100 babies/day) or a high anchor
(less than 50,000 babies/day). Those who saw the low anchor estimated many
fewer births/day than those who saw the high anchor, which suggests that the
wording can have a profound influence. The correct answer, as it happens, is
~11,000 births/day in 2014. To investigate the extent that these results are
replicable, the Many Labs project repeated this classic study at many
different labs around the world. You can find the summary data for 30 of
these labs in the Anchor Estimate ma data file
numerous other issues can be highly damaging, but are thriving in this social
media age. Trying to debunk a conspiracy theory by presenting facts and
evidence often doesn't work, alas. Psychological inoculation, also similar to
prebunking, presents a mild form of misinformation, preferably with
explanation, in the hope of building resistance to real-life fake news-a sort
of vaccine for fake news. The Bad News game is a spin-off from research on
psychological inoculation. Basol et al. (2020) assessed the possible
effectiveness of this game as a fake news vaccine. At getbadnews.com you can
click 'About' for information, or just start playing the game- it's easy and
maybe even fun. You encounter mock Twitter (now X) fake news messages that
illustrate common strategies for making fake news memorable or believable.
You make choices between messages and decide which ones to 'forward' as you
try to spread fake news while building your credibility score and number of
'followers'-rather like real life for a conspiracy theorist wanting to spread
the word. Compete with your friends for credibility and number of followers.
Basol's online participants first saw 18 fictitious fake news tweets and
rated each for reliability (accuracy, believability), and also rated their
confidence in that reliability rating. Both ratings were on a 1 to 7 scale.
Those in the BadNews group then played the game for 15 minutes, whereas those
in the Control group played Tetris. Then all once again gave reliability and
confidence ratings for the 18 tweets.
decade earlier, had been one of several outside experts invited to scrutinize
the laboratory and experimental procedures of parapsychology researcher
Charles Honorton. Bem not only judged them adequate, but joined the research
effort and became a coauthor with Honorton. Bem and Honorton (1994) first
reviewed early ganzfeld studies and described how the experimental procedure
had been improved to reduce the chance that results could be influenced by
various possible biases, or leakages of information from sender to receiver.
For example, the randomization procedure was carried out automatically by
computer, and all stimuli were presented under computer control. Bem and
Honorton then presented data from studies conducted with the improved
procedure. Table 13.1 presents basic data from 10 studies reported by Bem and
Honorton (1994). Participants each made a single judgment, so in Pilot 1, for
example, 22 participants responded, with 8 of them giving a correct response.
Three pilot studies helped refine the procedures, then four studies used
novice receivers. Study 5 used 20 students of music, drama, or dance as
receivers, in response to suggestions that creative people might be more
likely to show telepathy. Studies 6 and 7 used receivers who had participated
in an earlier study. The proportion of hits expected by chance is .25, and
Table 13.1 shows that all but Study 1 found proportions higher than .25.
who identified as female.  Data is Subjective Wellbeing abd Body Satisfaction.
who identified as male.  Data is Subjective Wellbeing abd Body Satisfaction.
Reported are measures of Subjective Wellbeing abd Body Satisfaction.
}
