% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_any.R
\name{meta_any}
\alias{meta_any}
\title{Estimate any meta effect}
\usage{
meta_any(
  data,
  yi,
  vi,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  effect_size_name = "Effect size",
  moderator_variable_name = "My moderator",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble with columns}

\item{yi}{yi}

\item{vi}{vi}

\item{labels}{labels}

\item{moderator}{logical if moderator has been passed}

\item{contrast}{contrast}

\item{effect_label}{effect_label}

\item{effect_size_name}{defaults to MDiff}

\item{moderator_variable_name}{mvn}

\item{random_effects}{logical}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_any} returns
}
