% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_d1.R
\name{meta_d1}
\alias{meta_d1}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_d1(
  data,
  ds,
  ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{ds}{comparison}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{el}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{n}{comparison}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_mean} returns
}
