% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_d2.R
\name{meta_d2}
\alias{meta_d2}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_d2(
  data,
  ds,
  comparison_ns,
  reference_ns,
  r = NULL,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  assume_equal_variance = TRUE,
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{ds}{comparison}

\item{comparison_ns}{comparison}

\item{reference_ns}{reference}

\item{r}{optional correlation between measures for w-s studies, NA otherwise}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{el}

\item{assume_equal_variance}{aev}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_d2} returns
}
