% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_mdiff_two.R
\name{meta_mdiff_two}
\alias{meta_mdiff_two}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_mdiff_two(
  data,
  comparison_means,
  comparison_sds,
  comparison_ns,
  reference_means,
  reference_sds,
  reference_ns,
  r = NULL,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  reported_effect_size = c("mean_difference", "smd_unbiased", "smd"),
  assume_equal_variance = FALSE,
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{comparison_means}{comparison}

\item{comparison_sds}{comparison}

\item{comparison_ns}{comparison}

\item{reference_means}{reference}

\item{reference_sds}{reference}

\item{reference_ns}{reference}

\item{r}{optional correlation between measures for w-s studies, NA otherwise}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{el}

\item{reported_effect_size}{smd}

\item{assume_equal_variance}{aev}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returns object of class esci_estimate
}
\description{
\code{meta_mdiff_two} returns
}
