% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_mean.R
\name{meta_mean}
\alias{meta_mean}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_mean(
  data,
  means,
  sds,
  ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  reference_mean = 0,
  reported_effect_size = c("mean_difference", "smd_unbiased", "smd"),
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{means}{comparison}

\item{sds}{comparison}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{el}

\item{reference_mean}{Reference mean, defaults to 0}

\item{reported_effect_size}{smd}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{n}{comparison}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_mean} returns
}
