% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_pdiff_two.R
\name{meta_pdiff_two}
\alias{meta_pdiff_two}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_pdiff_two(
  data,
  comparison_cases,
  comparison_ns,
  reference_cases,
  reference_ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  reported_effect_size = c("RD", "RR", "OR", "AS", "PETO"),
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{comparison_cases}{cases}

\item{comparison_ns}{comparison}

\item{reference_cases}{cases}

\item{reference_ns}{comparison}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{effect_label}

\item{reported_effect_size}{res}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_pdiff_two} returns
}
