% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_r.R
\name{meta_r}
\alias{meta_r}
\title{Estimate meta-analytic difference in magnitude between two ind. groups}
\usage{
meta_r(
  data,
  rs,
  ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{rs}{smd}

\item{ns}{comparison}

\item{labels}{labels}

\item{moderator}{mod}

\item{contrast}{contrast}

\item{effect_label}{effect_label}

\item{random_effects}{re}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\code{meta_r} returns
}
