% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{Calculates descriptive statistics for a continuous variable}
\usage{
overview(
  data = NULL,
  outcome_variable = NULL,
  grouping_variable = NULL,
  means = NULL,
  sds = NULL,
  ns = NULL,
  grouping_variable_levels = NULL,
  outcome_variable_name = "My Outcome Variable",
  grouping_variable_name = NULL,
  conf_level = 0.95,
  assume_equal_variance = FALSE
)
}
\arguments{
\item{data}{\itemize{
\item for raw data, a data frame or tibble
}}

\item{outcome_variable}{\itemize{
\item for raw data, either a vector containing numerical
data or the name of a data-frame column containing a factor
}}

\item{grouping_variable}{\itemize{
\item optional; for raw data either a vector containing
a factor or the name of a data frame column containing a factor
}}

\item{means}{For summary data - A vector of 1 or more numerical means}

\item{sds}{For summary data - A vector of standard deviations, same length as
means}

\item{ns}{For summary data - A vector of sample sizes, same length as means}

\item{grouping_variable_levels}{For summary data - An optional vector of
group labels, same length as means.  If not passed, auto-generated.}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My Outcome Variable'.  Ignored if a data-frame is passed,
this argument is ignored.}

\item{grouping_variable_name}{Optional friendly name for the grouping
variable.  If a data frame is passed, this argument is ignored.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{assume_equal_variance}{Defaults to FALSE}
}
\value{
Returns a table of descriptive statistics
}
\description{
This function calculates basic descriptive statistics for a numerical
variable.  It can calculate an overall summary, or broken down by
the levels of a grouping variable.  Inputs can be summary data,
vectors, or a data frame.
}
