% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Plot the interaction from a 2x2 design}
\usage{
plot_interaction(
  estimate,
  effect_size = c("mean", "median"),
  show_CI = FALSE,
  ggtheme = NULL,
  line_count = 100,
  line_alpha = 0.02
)
}
\arguments{
\item{estimate}{A esci_estimate object with raw data an es_mdiff_2x2_
function}

\item{effect_size}{Optional; one of 'mean' or 'median' to determine the
measure of central tendency plotted.  Note that median is only available if
the estimate was generated from raw data.  Defauls to 'mean'}

\item{show_CI}{Optional logical; set to TRUE to visualize the confidence
intervals on each simple effect; defaults to FALSE}

\item{ggtheme}{Optional ggplot2 theme object to specify the visual style of the
plot.  Defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}}

\item{line_count}{Optional integer > 0 to specify the number of lines used
to visualize the simple-effect confidence intervals; defaults to 100}

\item{line_alpha}{Optional numeric between 0 and 1 to specify the
alpha (transparancy) of the confidence interval lines; defaults to 0.02}
}
\description{
\code{plot_interaction} helps visualize the interaction from a 2x2 desing.
It plots the 2 simple effects for the first factor and can also
help visualize the CIs on those simple effects.  It is the comparison
between those simple effects that represents an interaction (the difference
in the difference).
You can pass esci-estimate objects generated
\code{\link[=estimate_mdiff_2x2_between]{estimate_mdiff_2x2_between()}} or \code{\link[=estimate_mdiff_2x2_mixed]{estimate_mdiff_2x2_mixed()}}.
This function returns a ggplot2 object.
}
