% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta.R
\name{plot_meta}
\alias{plot_meta}
\title{Generates a forest plot displaying results of a meta-analysis}
\usage{
plot_meta(
  estimate,
  mark_zero = TRUE,
  include_PIs = FALSE,
  report_CIs = FALSE,
  explain_DR = FALSE,
  meta_diamond_height = 0.35,
  ggtheme = ggplot2::theme_classic()
)
}
\arguments{
\item{estimate}{\itemize{
\item an esci_estimate object generated by an esci meta_ function
}}

\item{mark_zero}{\itemize{
\item Boolean; defaults to TRUE to include a dotted line
indicated no effect (effect_size = 0)
}}

\item{include_PIs}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to include
prediction intervals for the overall effect and each moderator level (if
defined)
}}

\item{report_CIs}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to include
printed representation of each study effect size and CI along the right-
hand of the figure
}}

\item{explain_DR}{\itemize{
\item Boolean; dedaults to FALSE; set to TRUE if no moderator
is defined to show both the RE and FE effect sizes to represent how the
diamond ration measure of effect-size heterogeneity is calculated
}}

\item{meta_diamond_height}{\itemize{
\item Optional real number > 0 to indicate height
that each meta-analytic diamond should be drawn; defaults to 0.35
}}

\item{ggtheme}{\itemize{
\item Optional ggplot2 theme object to control overall styling; defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}
}}
}
\description{
`plot_meta' returns a ggplot2 object visualizing the results of a meta-analysis, showing
each study effect size and CI, the overall effect size and CI as a diamond,
effect sizes estimated at each moderator level (if defined), and (optionally)
prediction intervals for subsequent studies.
}
\details{
This function requires as input an esci_estimate object generated by an
esci meta-analysis function: \code{\link[=meta_any]{meta_any()}}, \code{\link[=meta_d1]{meta_d1()}},
\code{\link[=meta_d2]{meta_d2()}}, \code{\link[=meta_mdiff_two]{meta_mdiff_two()}}, \code{\link[=meta_mean]{meta_mean()}}, \code{\link[=meta_pdiff_two]{meta_pdiff_two()}},
\code{\link[=meta_proportion]{meta_proportion()}}, and \code{\link[=meta_r]{meta_r()}}.
}
