% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generates a scatter plot of data for two continuous variables}
\usage{
plot_scatter(
  estimate,
  show_line = FALSE,
  show_line_CI = FALSE,
  show_PI = FALSE,
  show_residuals = FALSE,
  show_mean_lines = FALSE,
  show_r = FALSE,
  predict_from_x = NULL,
  plot_as_z = FALSE,
  ggtheme = ggplot2::theme_classic()
)
}
\arguments{
\item{estimate}{\itemize{
\item an esci_estimate object generated by \code{\link[=estimate_r]{estimate_r()}}
}}

\item{show_line}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show the
regression line
}}

\item{show_line_CI}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show the
confidence interval on the regression line
}}

\item{show_PI}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show prediction
intervals
}}

\item{show_residuals}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show
residuals of prediction
}}

\item{show_mean_lines}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to plot
lines showing the mean of each variable
}}

\item{show_r}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to print the \emph{r}
value and its CI on the plot
}}

\item{predict_from_x}{\itemize{
\item Optional real number in the range of the x variable
for the plot; Defaults to NULL; if passed, the graph shows the predicted Y'
for this x value
}}

\item{plot_as_z}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to convert x and y
scores to z scores prior to plotting
}}

\item{ggtheme}{\itemize{
\item Optional ggplot2 theme object to control overall styling;
defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}
}}
}
\value{
Returns an object of class esci_estimate
\itemize{
\item \strong{overview}
\itemize{
\item \emph{outcome_variable_name} -
\item \emph{mean} -
\item \emph{mean_LL} -
\item \emph{mean_UL} -
\item \emph{median} -
\item \emph{median_LL} -
\item \emph{median_UL} -
\item \emph{sd} -
\item \emph{min} -
\item \emph{max} -
\item \emph{q1} -
\item \emph{q3} -
\item \emph{n} -
\item \emph{missing} -
\item \emph{df} -
\item \emph{mean_SE} -
\item \emph{median_SE} -
}
\item \strong{es_mean}
\itemize{
\item \emph{outcome_variable_name} -
\item \emph{effect} -
\item \emph{effect_size} -
\item \emph{LL} -
\item \emph{UL} -
\item \emph{SE} -
\item \emph{df} -
\item \emph{ta_LL} -
\item \emph{ta_UL} -
}
\item \strong{es_median}
\itemize{
\item \emph{outcome_variable_name} -
\item \emph{effect} -
\item \emph{effect_size} -
\item \emph{LL} -
\item \emph{UL} -
\item \emph{SE} -
\item \emph{df} -
\item \emph{ta_LL} -
\item \emph{ta_UL} -
}
\item \strong{raw_data}
\itemize{
\item \emph{grouping_variable} -
\item \emph{outcome_variable} -
}
}
}
\description{
\code{plot_scatter} returns a ggplot2 object of data from two continuous
variables.  Can indicate regression line and its confidence
interval,prediction intervals regression residuals and more. This function
requires as input an esci_estimate object generated by \code{\link[=estimate_r]{estimate_r()}}
}
\details{
This function was developed primarily for student use within jamovi when
learning along with the text book Introduction to the New Statistics, 2nd
edition (Cumming & Calin-Jageman, 2024).

Expect breaking changes as this function is improved for general use.  Work
still do be done includes:
\itemize{
\item Revise to avoid deprecated ggplot features
\item Revise for consistent ability to control aesthetics and consistent
layer names
}
}
\examples{
# From raw data

thomason1 <- data.frame(
  ls_pre = c(
    13,
    12,
    12,
    9,
    14,
    17,
    14,
    9,
    6,
    7,
    11,
    15
  ),
  ls_post = c(
    14,
    13,
    16,
    12,
    15,
    18,
    13,
    10,
    10,
    8,
    14,
    16
  )
)

estimate <- esci::estimate_r(
  thomason1,
  ls_pre,
  ls_post
)
estimate

# To evaluate a hypothesis (interval null from -0.1 to 0.1):
test_correlation(estimate, rope = c(-0.1, 0.1))

\dontrun{
# To visualize the value of r
plot_correlation(estimate)

# To visualize the data (scatterplot)
plot_scatter(estimate)

# To visualize the data (scatterplot) and use regression to obtain Y' from X
plot_scatter(estimate, predict_from_x = 10)

}

# From summary data
estimate <- esci::estimate_r(r = 0.536, n = 50)
estimate


\dontrun{
# To visualize the estimate
plot_correlation(estimate)
}


}
