% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_estimate.R
\name{print.esci_estimate}
\alias{print.esci_estimate}
\title{Print an esci_estimate}
\usage{
\method{print}{esci_estimate}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{\itemize{
\item object to print; must be of class esci_estimate
}}

\item{...}{S3 signature for generic plot function.}

\item{verbose}{\itemize{
\item optional logical print all details; defaults to false
}}
}
\description{
Pretties up the printing of a complex esci_estimate object.
}
