% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_evaluation.R
\name{test_correlation}
\alias{test_correlation}
\title{Test a hypothesis about the strength of a Pearson's \emph{r} correlation}
\usage{
test_correlation(estimate, rope = c(0, 0), output_html = FALSE)
}
\arguments{
\item{estimate}{\itemize{
\item An esci_estimate object generated by the estimate_r
function
}}

\item{rope}{\itemize{
\item A two-element vector defining the Region of Practical
Equivalence (ROPE). Specify c(0, 0) to test a point null of exactly 0.
Specify any two ascending values between -1 and 1 to test an interval null
(e.g. c(.25, .45) to test the hypothesis that Pearson's \emph{r} in the
population (rho) is between .25 and .45).
}}

\item{output_html}{\itemize{
\item TRUE to return results in HTML; FALSE (default) to
return standard output
}}
}
\description{
\code{test_correlation} is suitable for testing a hypothesis about a
the strength of correlation between two continuous variables (designs
in which Pearson's \emph{r} is a suitable measure of correlation).
}
\details{
This function can be passed an esci_estimate object generated by
\code{\link[=estimate_r]{estimate_r()}}.

It can test hypotheses about a specific value for the difference (a
point null) or about a range of values (an interval null)
}
