% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_mdiff.R
\name{test_pdiff}
\alias{test_pdiff}
\title{Test a hypothesis about a difference in proportion}
\usage{
test_pdiff(estimate, rope = c(0, 0), output_html = FALSE)
}
\arguments{
\item{estimate}{\itemize{
\item An esci_estimate object generated by an estimate_pdiff_ function
}}

\item{rope}{\itemize{
\item A two-element vector defining the Region of Practical
Equivalence (ROPE). Specify c(0, 0) to test a point null of exactly 0.
Specify any two ascending values between -1 and 1 to test an interval null
(e.g. c(-.25, .25) to test the hypothesis that the difference in proportion
is between -.25 and .25).
}}

\item{output_html}{\itemize{
\item TRUE to return results in HTML; FALSE (default) to
return standard output
}}
}
\description{
\code{test_pdiff} is suitable for testing a hypothesis about a
difference in proportions between two conditions with a categorical outcome
variable.  It can test hypotheses for both independent and paired designs.
}
\details{
This function can be passed an esci_estimate object generated by
\code{\link[=estimate_pdiff_one]{estimate_pdiff_one()}}, \code{\link[=estimate_pdiff_two]{estimate_pdiff_two()}},
\code{\link[=estimate_pdiff_paired]{estimate_pdiff_paired()}}, or \code{\link[=estimate_pdiff_ind_contrast]{estimate_pdiff_ind_contrast()}}.

It can test hypotheses about a specific value for the difference (a
point null) or about a range of values (an interval null)
}
