% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_emotion_heartrate}
\alias{data_emotion_heartrate}
\title{Emotion heartrate - Ch8 - from Lakens (2013)}
\format{
\subsection{\code{data_emotion_heartrate}}{

A data frame with 68 rows and 3 columns:
\describe{
\item{ID}{integer}
\item{HR_baseline}{numeric}
\item{HR_anger}{numeric}
}
}
}
\source{
\doi{10.1109/T-AFFC.2013.3}
}
\usage{
data_emotion_heartrate
}
\description{
Anger is a powerful emotion. To what extent can feeling angry actually change
your heart rate? To investigate, Lakens (2013) asked students to record their
heart rate (in beats per minute) at rest before (baseline) and then while
recalling a time of intense anger. This is a conceptual replication of a
classic study by Ekman et al. (1983). Load the Emotion heartrate data set
from the book website.
}
\keyword{datasets}
