% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_flag_priming_ma}
\alias{data_flag_priming_ma}
\title{Flag Priming ma - Ch9 - Many Labs replications of Carter et al. (2011)}
\format{
\subsection{\code{data_flag_priming_ma}}{

A data frame with 25 rows and 7 columns:
\describe{
\item{Location}{factor}
\item{M Flag}{numeric}
\item{s Flag}{numeric}
\item{n Flag}{integer}
\item{M Noflag}{numeric}
\item{s Noflag}{numeric}
\item{n Noflag}{integer}
}
}
}
\source{
\doi{10.5334/jopd.ad}
}
\usage{
data_flag_priming_ma
}
\description{
To what extent does being exposed to the American flag influence political
attitudes? One seminal study (Carter et al., 2011) explored this issue by
subtly exposing participants either to images of the American flag or to
control images. Next, participants were asked about their political
attitudes, using a 1-7 rating scale where high scores indicate conservative
attitudes. Participants exposed to the flag were found to express
substantially more conservative attitudes. The Many Labs project replicated
this finding at 25 different locations in the United States.
}
\keyword{datasets}
