% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gender_math_iat}
\alias{data_gender_math_iat}
\title{Gender math IAT - Ch07 - Ithaca and SDSU replications of Nosek et al. (2002)}
\format{
\subsection{\code{data_gender_math_iat}}{

A data frame with 155 rows and 4 columns:
\describe{
\item{Ithaca gender}{factor}
\item{Ithaca IAT}{numeric}
\item{SDSU gender}{factor}
\item{SDSU IAT}{numeric}
}
}
}
\source{
\doi{10.1027/1864-9335/a000178}
}
\usage{
data_gender_math_iat
}
\description{
To what extent do men and women differ in their attitudes towards
mathematics? To investigate, Nosek et al. (2002) asked male and female
students to complete an Implicit Association Test (IAT)-this is a task
designed to measure a participant's implicit (non-conscious) feelings towards
a topic. (If you've never heard of the IAT, try it out here:
tiny.cc/harvardiat) On this IAT, students were tested for negative feelings
towards mathematics and art. Scores reflect the degree to which a student had
more negative implicit attitudes about mathematics than art (positive score:
more negative feelings about mathematics; 0: same level of negativity to
both; negative score: more negative feelings about art). data_gender_math_iat
has data from two labs that participated in a large-scale replication of the
original study (Klein et al., 2014a, 2014b)
}
\keyword{datasets}
