% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gender_math_iat_ma}
\alias{data_gender_math_iat_ma}
\title{Gender math IAT ma - Ch9 - Many Labs replications of Nosek et al. (2002)}
\format{
\subsection{\code{data_gender_math_iat_ma}}{

A data frame with 30 rows and 9 columns:
\describe{
\item{Location}{factor}
\item{M Male}{numeric}
\item{s Male}{numeric}
\item{n Male}{integer}
\item{M Female}{numeric}
\item{s Female}{numeric}
\item{n Female}{integer}
\item{USAorNot}{factor}
\item{Country}{factor}
}
}
}
\source{
\doi{10.1027/1864-9335/a000178}
}
\usage{
data_gender_math_iat_ma
}
\description{
In EOC Exercise 4 in Chapter 7 we encountered the classic study of Nosek et
al. (2002), in which male and female participants completed an Implicit
Association Test (IAT) that measured the extent of negative attitudes towards
mathematics, compared with art. The study found that women, compared with
men, tended to have more negative implicit attitudes towards mathematics. The
Many Labs project repeated this study at locations around the world (Klein et
al., 2014a, 2014b). Summary data for 30 of these labs are available in Gender
math IAT ma. Higher scores indicate more implicit bias against mathematics.
See also data_gender_math_iat for raw data from two specific sites from this
replication effort.
}
\keyword{datasets}
