% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_halagappa}
\alias{data_halagappa}
\title{Halagappa - Ch14 - from Halagappa et al. (2007)}
\format{
\subsection{\code{data_halagappa}}{

A data frame with 6 rows and 4 columns:
\describe{
\item{Groups}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{n}{integer}
}
}
}
\source{
\doi{10.1016/j.nbd.2006.12.019}
}
\usage{
data_halagappa
}
\description{
Could eating much less delay Alzheimer's? If so, that would be great news.
Halagappa et al. (2007) investigated the possibility by using a mouse model,
meaning they used Alzheimer-prone mice, which were genetically predisposed to
develop neural degeneration typical of Alzheimer's. The researchers used six
independent groups of mice, three tested in mouse middle age when 10 months
old, and three in mouse old age when 17 months. At each age there was a
control group of normal mice that ate freely (the NFree10 and NFree17
groups), a group of Alzheimer-prone mice that also ate freely (the AFree10
and AFree17 groups), and another Alzheimer-prone group restricted to 40\% less
food than normal (the ADiet10 and ADiet17 groups). Table 14.2 lists the
factors that define the groups, and group labels. I'll discuss one measure of
mouse cognition: the percent time spent near the target of a water maze, with
higher values indicating better learning and memory. Table 14.2 reports the
means and standard deviations for this measure, and group sizes.
}
\keyword{datasets}
