% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_home_prices}
\alias{data_home_prices}
\title{Home Prices - Ch12 - for End-of-Chapter Exercise 12.2}
\format{
\subsection{\code{data_home_prices}}{

A data frame with 300 rows and 8 columns:
\describe{
\item{MLS}{integer}
\item{Location}{factor}
\item{Price}{integer}
\item{Bedrooms}{integer}
\item{Bathrooms}{integer}
\item{Size (m2)}{numeric}
\item{Status}{factor}
\item{Status_Code}{logical}
}
}
}
\usage{
data_home_prices
}
\description{
Maybe you're thinking about buying a house after college? Regression can help
you hunt for a bargain. Download the Home Prices data set. This file contains
real estate listings from 1997 to 2003 in a city in California. Let's explore
the extent to which the size of the home (in square meters) predicts the sale
price.
}
\keyword{datasets}
