% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_labels_flavor}
\alias{data_labels_flavor}
\title{Labels flavor - Ch8 - from Floretta-Schiller et al. (2015)}
\format{
\subsection{\code{data_labels_flavor}}{

A data frame with 51 rows and 6 columns:
\describe{
\item{ParticipantID}{integer}
\item{Enjoy_Generic}{numeric}
\item{Enjoy_Organic}{numeric}
\item{Pay_Generic}{numeric}
\item{Pay_Organic}{numeric}
\item{Suspicious}{logical}
}
}
}
\usage{
data_labels_flavor
}
\description{
To what extent do brand labels influence perceptions of a product? To
investigate, participants were asked to participate in a taste test. All
participants were actually given the same grape juice, but one glass was
poured from a bottle labeled 'Organic' and the other glass from a bottle
labeled 'Generic'. After each tasting (in counterbalanced order),
participants were asked to rate how much they enjoyed the juice on a scale
from 1 (not at all) to 10 (very much). Participants were also asked to say
how much they'd be willing to pay for a large container of that juice on a
scale from $1 to $10. Load the Labels flavor data set from the book website.
These data were collected as part of a class project by Floretta-Schiller et
al. (2015), whose work was inspired by a very clever study looking at the
effects of fast-food wrappers on children's enjoyment of food (Robinson et
al., 2007).
}
\keyword{datasets}
