% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_latimier_3groups}
\alias{data_latimier_3groups}
\title{Latimier 3Groups - Ch14 - 3 groups in Latimier et al. (2019)}
\format{
\subsection{\code{data_latimier_3groups}}{

A data frame with 285 rows and 3 columns:
\describe{
\item{Group}{factor}
\item{Test\%}{numeric}
\item{Time}{numeric}
}
}
}
\source{
\doi{10.7910/DVN/XPYPMF}
}
\usage{
data_latimier_3groups
}
\description{
The researchers were interested in how different study approaches might
impact learning. Working in France, they created three independent groups,
each comprising 95 adults. Participants worked online through seven learning
modules about DNA. The Reread group worked through a module, then worked
through it a second time before going on to the next module. The Quiz group
worked through a module, then had to complete a quiz before going on to the
next module. The Prequiz group had to work through the quiz before seeing the
presentation of a module, then went on to the quiz and presentation of the
next module. Participants received feedback and a brief explanation after
answering each question in a quiz, and could take as long as they wished to
work through each module and quiz. Seven days later, participants completed a
final test.
data_latimier_3groups is the full data set.
To facilitate different student exercises, there are also separate data
entities for each group (data_latimier_prequiz, data_latimier_reread, etc.),
and for every \emph{pair} of groups (data_latimier_quiz_prequiz, etc.).
}
\keyword{datasets}
