% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_latimier_quiz_prequiz}
\alias{data_latimier_quiz_prequiz}
\title{Latimier Quiz Prequiz - Ch07 - Quiz and Prequiz groups in Latimier et al. (2019)}
\format{
\subsection{\code{data_latimier_quiz_prequiz}}{

A data frame with 190 rows and 3 columns:
\describe{
\item{Group}{factor}
\item{Test\%}{numeric}
\item{Time}{numeric}
}
}
}
\source{
\doi{10.7910/DVN/XPYPMF}
}
\usage{
data_latimier_quiz_prequiz
}
\description{
Just the Quiz (RQ) an Prequiz (QR) groups from Latimier et al., 2019
See full details in data_latimier_3_groups
}
\keyword{datasets}
