% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_latimier_reread}
\alias{data_latimier_reread}
\title{Latimier Reread - Ch03 - Reread group in Latimier et al. (2019)}
\format{
\subsection{\code{data_latimier_reread}}{

A data frame with 95 rows and 3 columns:
\describe{
\item{Group}{factor}
\item{Test\%}{numeric}
\item{Time}{numeric}
}
}
}
\source{
\doi{10.7910/DVN/XPYPMF}
}
\usage{
data_latimier_reread
}
\description{
Just the Reread group from Latimier et al., 2019
See full details in data_latimier_3_groups
}
\keyword{datasets}
