% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_organicmoral}
\alias{data_organicmoral}
\title{OrganicMoral - Ch14 - from Eskine (2013)}
\format{
\subsection{\code{data_organicmoral}}{

A data frame with 106 rows and 6 columns:
\describe{
\item{Group}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{N}{integer}
\item{ReplicationGroup}{factor}
\item{MoralJudgmentment}{numeric}
}
}
}
\source{
\doi{10.1177/1948550616639649}
}
\usage{
data_organicmoral
}
\description{
To what extent might choosing organic foods make us morally smug? To
investigate, Eskine (2013) asked participants to rate images of organic food,
neutral (control) food, or comfort food. Next, under the guise of a different
study, all participants completed a moral judgment scale in which they read
different controversial scenarios and rated how morally wrong they judged
them to be (scale of 1-7, high judgments mean more wrong). Table 14.7 shows
summary data, which are also available in the first four variables in the
OrganicMoral file. In that file you can see two further variables, which
report full data-we'll come to these shortly. Here we use the summary data.
After the results of Eskine (2013) were published, Moery and Calin-Jageman
(2016) conducted a series of close replications. We obtained original
materials from Eskine, piloted the procedure, and preregistered our sampling
and analysis plan. The OSF page, osf.io/atkn7, has all the details. The data
from one of these close replications are in the last two variables of the
OrganicMoral file. For this replication study, group names are in the
variable ReplicationGroup and moral judgments in MoralJudgment. (You may need
to scroll right to see these variables.)
}
\keyword{datasets}
