% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_penlaptop1}
\alias{data_penlaptop1}
\title{\% transcription scores from pen and laptop group of Meuller et al., 2014}
\format{
\subsection{\code{data_penlaptop1}}{

A data frame with 65 rows and 2 columns:
\describe{
\item{condition}{factor}
\item{transcription}{numeric}
}
}
}
\source{
\doi{10.1177/0956797614524581}
}
\usage{
data_penlaptop1
}
\description{
\% transcription scores from pen and laptop group of Meuller et al., 2014
}
\keyword{datasets}
