% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_powerperformance_ma}
\alias{data_powerperformance_ma}
\title{PowerPerformance ma - Ch9 - from Burgmer and Englich (2012), and Cusack et al. (2015)}
\format{
\subsection{\code{data_powerperformance_ma}}{

A data frame with 8 rows and 12 columns:
\describe{
\item{StudyName}{factor}
\item{Country}{factor}
\item{Population}{factor}
\item{Difficulty}{factor}
\item{Task}{factor}
\item{M Control}{numeric}
\item{s Control}{numeric}
\item{M Power}{numeric}
\item{s Power}{numeric}
\item{Cohen d unb}{numeric}
\item{n Control}{integer}
\item{n Power}{integer}
}
}
}
\source{
\doi{10.1371/journal.pone.0140806}
}
\usage{
data_powerperformance_ma
}
\description{
To what extent could feeling powerful affect your performance at motor
skills? To investigate, Burgmer and Englich (2012) assigned German
participants to either power or control conditions and then asked them to
play golf (Experiment 1) or darts (Experiment 2). They found that
participants manipulated to feel powerful performed substantially better than
those in the control condition. To study this finding further, Cusack et al.
(2015) conducted five replications in the United States. Across these
replications they tried different ways of manipulating power, different types
of tasks (golf, mirror tracing, and a cognitive task), different levels of
difficulty, and different types of participant pools (undergraduates and
online). Summary data from all seven studies are available in
PowerPerformance ma.
}
\keyword{datasets}
