% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_religionsharing}
\alias{data_religionsharing}
\title{ReligionSharing - Ch14 - \strong{RETRACTED DATA} used in End-of-Chapter Exercise 14.3}
\format{
\subsection{\code{data_religionsharing}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{Group}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{N}{integer}
}
}
}
\source{
\doi{10.1016/j.cub.2015.09.056}
}
\usage{
data_religionsharing
}
\description{
To what extent is a religious upbringing related to prosocial behavior in
childhood? To investigate, a large international sample of children was asked
to play a game in which they were given 10 stickers but then asked if they
would give some of these stickers away to another child who had not been able
to be tested that day. The number of stickers donated was considered a
measure of altruistic sharing. In addition, the parents of each child
reported the family's religion. Summary data provided.  \strong{THIS STUDY
HAS BEEN RETRACTED DUE TO AN ERRONEOUS ANALYSIS - THE DATASET WILL BE
REMOVED FROM FUTURE VERSIONS OF ESCI AND THE BOOK}
}
\keyword{datasets}
