% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_simmonscredibility}
\alias{data_simmonscredibility}
\title{SimmonsCredibility - Ch14 - from Simmons and Nelson (2020)}
\format{
\subsection{\code{data_simmonscredibility}}{

A data frame with 3 rows and 4 columns:
\describe{
\item{Groups}{factor}
\item{Mean}{numeric}
\item{SD}{numeric}
\item{n}{integer}
}
}
}
\source{
\url{http://datacolada.org/94}
}
\usage{
data_simmonscredibility
}
\description{
You're excited! Your company has developed a wonderful new weight-loss
program, and now it's your job to develop the ad campaign. Should you choose
a BeforeAfter pair of pictures, as in Figure 14.1, top panel? Or might a
Progressive sequence of pictures of the same person, as in the bottom panel,
be more effective? Pause, think, and discuss. Which would you choose, and
why? You might think that BeforeAfter is simpler and more dramatic. On the
other hand, Progressive highlights the steady improvement that you claim the
program will deliver. You're probably not surprised to learn that BeforeAfter
is used often and has long been a favorite of the advertising industry,
whereas Progressive is used only rarely. Luca Cian and colleagues (Cian et
al., 2020) were curious to know the extent to which BeforeAfter is actually
more effective, appealing, and credible than Progressive, or, indeed, whether
Progressive might score more highly. They reported seven studies of various
aspects of that question. I'll focus on their Study 2, in which they used
three independent groups to compare all three conditions illustrated in
Figure 14.1. The BeforeAfterInfo condition, in the middle panel, comprises
three BeforeAfter pairs, thus providing extra information about the before
and after endpoints. The researchers included this condition in case any
advantage of Progressive might stem simply from having more images, rather
than because it illustrates a clear progressive sequence. They randomly
assigned 213 participants from MTurk to one of the three groups. Participants
were asked to 'imagine that you have decided to lose some weight', then saw
one of the three ads for a weight loss program called MRMDiets. They then
answered the question 'How would you evaluate MRMDiets?' by choosing a 1-7
response on several scales, including Unlikeable-Likable,
Ineffective-Effective, and Not credible-Credible. The researchers averaged
six such scores to give an overall Credibility score, on the 1-7 scale, with
7 being the most credible. Simmons and Nelson (2020) were sufficiently
intrigued to carry out two substantial very close replications. With the
cooperation of the original researchers, they used the same materials and
procedure. They used much larger groups and preregistered their research
plan, including data analysis plan. I'll focus on their first replication, in
which 761 participants from MTurk were randomized to the three groups.
}
\keyword{datasets}
