% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sleep_beauty}
\alias{data_sleep_beauty}
\title{Sleep Beauty - Ch11 - for End-of-Chapter Exercise 11.6}
\format{
\subsection{\code{data_sleep_beauty}}{

A data frame with 70 rows and 2 columns:
\describe{
\item{Sleep (hours)}{numeric}
\item{Rated_Attractiveness}{numeric}
}
}
}
\usage{
data_sleep_beauty
}
\description{
Is there really such a thing as beauty sleep? To investigate, researchers
decided to examine the extent to which sleep relates to attractiveness. Each
of 70 college students self-reported the amount of sleep they had the night
before. In addition, a photograph was taken of each participant and rated for
attractiveness on a scale from 1 to 10 by two judges of the opposite gender.
The average rating score was used. You can download this data set (Sleep
Beauty) from the book website.
}
\keyword{datasets}
