% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_stickgold}
\alias{data_stickgold}
\title{Stickgold - Ch06 - from Stickgold et al. (2000)}
\format{
\subsection{\code{data_stickgold}}{

A data frame with 11 rows and 3 columns:
\describe{
\item{Sleep deprived}{numeric}
\item{B}{factor}
\item{C}{factor}
}
}
}
\source{
\url{https://www.statcrunch.com/app/index.html?dataid=1053539}
}
\usage{
data_stickgold
}
\description{
Stickgold et al. (2000) found that, remarkably, performance on a visual
discrimination task actually improved over the 48-96 hours after initial
training, even without practice during that time. However, what if
participants were sleep deprived during that period? They trained 11
participants in that new skill, then all were sleep deprived. The data were
(-14.7, -10.7, -10.7, 2.2, 2.4, 4.5, 7.2, 9.6, 10, 21.3, 21.8)-or download
the Stickgold data set from the book website. The data are the changes in
performance scores from immediately after training to after the night without
sleep: 0 represents no change, positive scores represent improvement, and
negative scores represent decline. Data set courtesy of
DataCrunch (tiny.cc/Stickgold)
}
\keyword{datasets}
