% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_studystrategies}
\alias{data_studystrategies}
\title{StudyStrategies - Ch14 - from O'Reilly et al. (1998)}
\format{
\subsection{\code{data_studystrategies}}{

A data frame with 3 rows and 10 columns:
\describe{
\item{Group}{factor}
\item{TestMean}{numeric}
\item{TestSD}{numeric}
\item{TestN}{integer}
\item{PrevKnowMean}{numeric}
\item{PrevKnowSD}{numeric}
\item{PrevKnowN}{integer}
\item{EaseUseMean}{numeric}
\item{EaseUseSD}{numeric}
\item{EaseUseN}{integer}
}
}
}
\source{
\doi{10.1006/ceps.1997.0977}
}
\usage{
data_studystrategies
}
\description{
To what extent does study strategy influence learning? To investigate,
psychology students were randomly assigned to three groups and asked to learn
biology facts using one of three different strategies: a) Self-Explain
(explaining for each fact what new knowledge is gained and how it relates to
what is already known), b) Elab Interrogation (elaborative interrogation:
stating for each fact why it makes sense), or c) Repetition Control (stating
each fact over and over). After studying, students took a 25-point
fill-the-blank test (O'Reilly et al., 1998)
}
\keyword{datasets}
