% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_thomason_1}
\alias{data_thomason_1}
\title{Thomason 1 - Ch11 - from Thomason 1}
\format{
\subsection{\code{data_thomason_1}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{Participant ID}{factor}
\item{Pretest}{numeric}
\item{Posttest}{numeric}
}
}
}
\usage{
data_thomason_1
}
\description{
Summary data from an unpublished study by Neil Thomason and colleagues, who
were interested in ways to enhance students' critical thinking. They were
investigating argument mapping, which is a promising way to use diagrams to
represent the structure of arguments. Students in their study completed an
established test of critical thinking (the Pretest), then a critical thinking
course based on argument mapping, then a second version of the test (the
Posttest).
}
\keyword{datasets}
