% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_videogameaggression}
\alias{data_videogameaggression}
\title{VideogameAggression - Ch15 - from Hilgard (2015)}
\format{
\subsection{\code{data_videogameaggression}}{

A data frame with 223 rows and 3 columns:
\describe{
\item{Violence}{factor}
\item{Difficulty}{factor}
\item{Agression}{numeric}
}
}
}
\source{
\doi{10.1177/0956797619829688}
}
\usage{
data_videogameaggression
}
\description{
Video games can be violent and they can also be challenging. To what extent
might these factors cause aggressive behavior? To explore, Hilgard (2015)
asked male participants to play one of four versions of a video game for 15
minutes. The game was customized so that it could vary in violence (shooting
zombies or helping aliens) and difficulty (targets controlled by tough AI or
dumb AI). After the game, players were provoked by being given an insulting
evaluation by a confederate. Participants then got to decide how long the
confederate should hold their hand in painfully cold ice water (between 0 and
80 seconds), and this was taken as a measure of aggressive behavior. You can
find the materials and analysis plan for this study on the Open Science
Framework: osf. io/cwenz. This is a simplified version of the full data set.
}
\keyword{datasets}
