% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_r.R
\name{estimate_correlation}
\alias{estimate_correlation}
\title{Estimates the linear correlation (Pearson's r) between two continuous
variables}
\usage{
estimate_correlation(
  data = NULL,
  x = NULL,
  y = NULL,
  r = NULL,
  n = NULL,
  x_variable_name = "My x variable",
  y_variable_name = "My y variable",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - A data frame or tibble}

\item{x}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{y}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{r}{For summary data - A pearson's r correlation coefficient}

\item{n}{For summary data - Sample size, an integer > 0}

\item{x_variable_name}{Optional friendly name for the x variable.
Defaults to 'My x variable' or the outcome variable column name if a
data frame is passed.}

\item{y_variable_name}{Optional friendly name for the y variable.
Defaults to 'My y variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returns object of class esci_estimate
}
\description{
\code{estimate_correlation} is suitable for a design with two continuous
variables.  It estimates the linear correlation between two variables
(Pearson's r).  You can pass raw data or summary data.
}
\details{
Once you generate an estimate with this function, you can visualize
it with \code{\link[=plot_correlation]{plot_correlation()}} and you can test hypotheses with
\code{\link[=test_correlation]{test_correlation()}}.  In addition, you can use \code{\link[=plot_scatter]{plot_scatter()}}
to visualize the raw data and to conduct a
regression analysis that returns predicted Y' values from a given X value.

The estimated correlation is from \code{\link[statpsych:ci.cor]{statpsych::ci.cor()}}, which uses the
Fisher r-to-z approach.
}
