% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_magnitude.R
\name{estimate_magnitude}
\alias{estimate_magnitude}
\title{Estimates for a continuous variable with no grouping (single-group design)}
\usage{
estimate_magnitude(
  data = NULL,
  outcome_variable = NULL,
  mean = NULL,
  sd = NULL,
  n = NULL,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - A data frame or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{mean}{For summary data - A numeric representing the mean of the outcome
variable}

\item{sd}{For summary data - A numeric > 0, standard deviation of the outcome
variable}

\item{n}{For summary data - An integer > 0, sample size of the outcome
variable}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returns an object of class esci_estimate
}
\description{
\code{estimate_magnitude} is suitable for a single group design with a
continuous outcome variable.  It estimates the population mean and
population median (raw data only) with confidence intervals.  You can
pass raw data or summary data.
}
\details{
Reach for this function in place of a one-sample \emph{t}-test or \emph{z}-test.

Once you generate an estimate with this function, you can visualize
it with \code{\link[=plot_magnitude]{plot_magnitude()}}.

If you want to compare your sample to a known value or reference, then
use \code{\link[=estimate_mdiff_one]{estimate_mdiff_one()}}.

The estimated mean is from \code{\link[statpsych:ci.mean1]{statpsych::ci.mean1()}}.

The estimated median is from \code{\link[statpsych:ci.median1]{statpsych::ci.median1()}}
}
