% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mdiff_2x2_mixed.R
\name{estimate_mdiff_2x2_mixed}
\alias{estimate_mdiff_2x2_mixed}
\title{Estimate the magnitude of difference for an independent groups contrast.}
\usage{
estimate_mdiff_2x2_mixed(
  data,
  outcome_variable_level1,
  outcome_variable_level2,
  grouping_variable,
  outcome_variable_name = "My outcome variable",
  repeated_measures_name = "Time",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable_level1}{For raw data - The column name of the outcome
variable for level 1}

\item{outcome_variable_level2}{For raw data - The column name of the outcome
variable for level 2}

\item{grouping_variable}{For raw data - The column name of the grouping
variable; only 2 levels allowed}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{repeated_measures_name}{Optional friendly name for the repeated
measures name.  Defaults to 'Time'}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_mdiff_2x2_mixed} returns point estimate and
confidence intervals for a 2x2 mixed design
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_mdiff_2x2_between(

)

}
