% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mdiff_ind_contrast.R
\name{estimate_mdiff_ind_contrast}
\alias{estimate_mdiff_ind_contrast}
\title{Estimates for a multi-group design with a continuous outcome
variable}
\usage{
estimate_mdiff_ind_contrast(
  data = NULL,
  outcome_variable = NULL,
  grouping_variable = NULL,
  means = NULL,
  sds = NULL,
  ns = NULL,
  contrast = NULL,
  grouping_variable_levels = NULL,
  outcome_variable_name = "My outcome variable",
  grouping_variable_name = "My grouping variable",
  conf_level = 0.95,
  assume_equal_variance = FALSE,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - a data frame or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{grouping_variable}{For raw data - The column name of the grouping
variable, or a vector of group names}

\item{means}{For summary data - A vector of 2 or more means}

\item{sds}{For summary data - A vector of standard deviations, same length as
means}

\item{ns}{For summary data - A vector of sample sizes, same length as means}

\item{contrast}{A vector of group weights, same length as number of groups.}

\item{grouping_variable_levels}{For summary data - An optional vector of
group labels, same length as means}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{grouping_variable_name}{Optional friendly name for the grouping
variable.  Defaults to 'My grouping variable' or the grouping variable
column name if a data.frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{assume_equal_variance}{Defaults to FALSE}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returnsobject of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_mdiff_ind_contrast} is suitable for a multi-group design
(between subjects) with a continuous outcome variable.  It accepts
a user-defined set of contrast weights that allows estimation of any
1-df contrast.  It can express estimates as mean differences, standardized
mean differences (Cohen's d) or median differences (raw data only).  You can
pass raw data or summary data.
}
\details{
Reach for this function in place of a one-way ANOVA.

Once you generate an estimate with this function, you can visualize
it with \code{\link[=plot_mdiff]{plot_mdiff()}} and you can test hypotheses with
\code{\link[=test_mdiff]{test_mdiff()}}.

The estimated mean differences are from \code{\link[statpsych:ci.lc.mean.bs]{statpsych::ci.lc.mean.bs()}}.

The estimated SMDs are from \code{\link[=CI_smd_ind_contrast]{CI_smd_ind_contrast()}} which relies
on \code{\link[statpsych:ci.lc.stdmean.bs]{statpsych::ci.lc.stdmean.bs()}} unless there are only 2 groups.

The estimated median differences are from \code{\link[statpsych:ci.lc.median.bs]{statpsych::ci.lc.median.bs()}}
}
