% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mdiff_paired.R
\name{estimate_mdiff_paired}
\alias{estimate_mdiff_paired}
\title{Estimate magnitude of difference between two related measures.}
\usage{
estimate_mdiff_paired(
  data = NULL,
  comparison_measure = NULL,
  reference_measure = NULL,
  comparison_mean = NULL,
  comparison_sd = NULL,
  reference_mean = NULL,
  reference_sd = NULL,
  n = NULL,
  correlation = NULL,
  comparison_measure_name = "Comparison measure",
  reference_measure_name = "Reference measure",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{comparison_measure}{For raw data - The column name of comparison
measure of the outcome variable, or a vector of numeric data}

\item{reference_measure}{For raw data - The column name of the reference
measure of the outcome variable, or a vector of numeric data}

\item{comparison_mean}{For summary data, a numeric}

\item{comparison_sd}{For summary data, numeric > 0}

\item{reference_mean}{For summary data, a numeric}

\item{reference_sd}{For summary data, numeric > 0}

\item{n}{For summary data, a numeric integer > 0}

\item{correlation}{For summary data, correlation between measures, numeric > -1 and < 1}

\item{comparison_measure_name}{For summary data - An optional character
label for the comparison measure.  Defaults to 'Comparison measure'}

\item{reference_measure_name}{For summary data - An optional character
label for the reference measure.  Defaults to 'Reference measure'}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returns object of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_mdiff_paired} returns effect sizes estimating the magnitude of
difference between two related quantiative measures.
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_mdiff_paired(
 datasets::iris,
 Sepal.Length,
 Petal.Length
)

}
