% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pdiff_one.R
\name{estimate_pdiff_one}
\alias{estimate_pdiff_one}
\title{Estimates for a single-group design with a categorical outcome
variable compared to a reference or population value.}
\usage{
estimate_pdiff_one(
  data = NULL,
  outcome_variable = NULL,
  comparison_cases = NULL,
  comparison_n = NULL,
  reference_p = 0,
  case_label = 1,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  count_NA = FALSE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{outcome_variable}{For raw data - The column name of the outcome
variable, which must be a factor, or a vector that is a factor}

\item{comparison_cases}{For summary data, a numeric integer > 0}

\item{comparison_n}{For summary data, a numeric integer >= count}

\item{reference_p}{Reference proportion, numeric >=0 and <=1}

\item{case_label}{An optional numeric or character label for the
count level.}

\item{outcome_variable_name}{Optional friendly name for the outcome variable.
Defaults to 'My outcome variable' or the outcome variable column name if a
data frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{count_NA}{Logical to count NAs (TRUE) in total N or not (FALSE)}
}
\value{
Returns an object of class esci_estimate
}
\description{
\loadmathjax
\code{estimate_pdiff_one} is suitable for a single-group design
(between subjects) with a categorical outcome variable.  It calculates
the effect sizes with respect to a reference or population proportion
(default value of 0).  It returns the estimated difference between the
in proportion from the reference/population value.  You can pass raw data or
summary data.
}
\details{
Once you generate an estimate with this function, you can visualize
it with \code{\link[=plot_pdiff]{plot_pdiff()}} and you can test hypotheses with
\code{\link[=test_pdiff]{test_pdiff()}}.

The estimated proportion differences are from \code{\link[statpsych:ci.prop1]{statpsych::ci.prop1()}}.
}
