% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rdiff_two.R
\name{estimate_rdiff_two}
\alias{estimate_rdiff_two}
\title{Estimate a correlation.}
\usage{
estimate_rdiff_two(
  data = NULL,
  x = NULL,
  y = NULL,
  grouping_variable = NULL,
  comparison_r = NULL,
  comparison_n = NULL,
  reference_r = NULL,
  reference_n = NULL,
  grouping_variable_levels = NULL,
  x_variable_name = "My x variable",
  y_variable_name = "My y variable",
  grouping_variable_name = "My grouping variable",
  conf_level = 0.95,
  save_raw_data = TRUE
)
}
\arguments{
\item{data}{For raw data - a dataframe or tibble}

\item{x}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{y}{For raw data - The column name of the outcome
variable, or a vector of numeric data}

\item{grouping_variable}{For raw data, a vector that is a factor or the
name of a factor column from data}

\item{comparison_r}{For summary data, a pearson's r correlation coefficient}

\item{comparison_n}{For summary data - An integer > 0}

\item{reference_r}{For summary data, a pearson's r correlation coefficient}

\item{reference_n}{For summary data - An integer > 0}

\item{grouping_variable_levels}{For summary data - An optional vector of
2 group labels}

\item{x_variable_name}{Optional friendly name for the x variable.
Defaults to 'My x variable' or the outcome variable column name if a
data frame is passed.}

\item{y_variable_name}{Optional friendly name for the y variable.
Defaults to 'My y variable' or the outcome variable column name if a
data frame is passed.}

\item{grouping_variable_name}{Optional friendly name for the grouping
variable.  Defaults to 'My grouping variable' or the grouping variable
column name if a data.frame is passed.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}

\item{save_raw_data}{For raw data; defaults to TRUE; set to FALSE to save
memory by not returning raw data in estimate object}
}
\value{
Returnsobject of class esci_estimate
}
\description{
Returns effect sizes appropriate for estimating the linear relationship
between two quantitative variables
}
\examples{
# From Raw Data ------------------------------------
# Just pass in the data source, grouping column, and outcome column.
# You can pass these in by position, skipping the labels:

# Note... not sure if PlantGrowth dataset meets assumptions for this analysis
estimate_correlation(
 datasets::anscombe,
 x1,
 y1
)

}
