% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_meta_diamond.R
\docType{data}
\name{Geom_meta_diamond_h}
\alias{Geom_meta_diamond_h}
\title{Create a new ggproto object}
\format{
An object of class \code{Geom_meta_diamond_h} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
Geom_meta_diamond_h
}
\arguments{
\item{_class}{Class name to assign to the object. This is stored as the class
attribute of the object. This is optional: if \code{NULL} (the default),
no class name will be added to the object.}

\item{_inherit}{ggproto object to inherit from. If \code{NULL}, don't
inherit from any object.}

\item{...}{A list of members in the ggproto object.}
}
\description{
Create a new ggproto object
}
\keyword{datasets}
