% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_any.R
\name{meta_any}
\alias{meta_any}
\title{Estimate any meta effect.}
\usage{
meta_any(
  data,
  yi,
  vi,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  effect_size_name = "Effect size",
  moderator_variable_name = "My moderator",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A data frame or tibble with columns}

\item{yi}{Name a column in data containing the effect size for each study}

\item{vi}{Name of a column in data containing the expected sampling variance
for each study}

\item{labels}{Name of a column in data containing a label for each study}

\item{moderator}{Optional name of a column in data containing a factor as a
categorical moderator}

\item{contrast}{Optional vector specifying a contrast analysis for the
categorical moderator.  Only define if a moderator is defined; vector
length should match number of levels in the moderator}

\item{effect_label}{Optional human-friendly name for the effect being
synthesized; defaults to 'My effect'}

\item{effect_size_name}{Optional human-friendly name of the effect size being
synthesized; defaults to 'Effect size'}

\item{moderator_variable_name}{Optional human-friendly name of the moderator,
if defined; If not passed but a moderator is defined, will be set to the
quoted name of the moderator column or 'My moderator'}

\item{random_effects}{Use TRUE to obtain a random effect meta-anlaysis
(usually reccomended); FALSE for fixed effect.}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
An esci-estimate object; a list of data frames and properties.
Returned tables include:
\itemize{
\item es_meta - A data frame with one overall.  If a moderator was
defined, there is an additional row for each level of the moderator.
\itemize{
\item effect_label -  Study label
\item effect_size - Effect size
\item LL - Lower bound of conf_level\% confidence interval
\item UL - Upper bound of conf_level\% confidence interval
\item SE - Expected standard error
\item k - Number of studies
\item diamond_ratio - ratio of random to fixed effects meta-analtics effect sizes
\item diamond_ratio_LL - lower bound of conf_level\% confidence interval for diamond ratio
\item diamond_ratio_UL - upper bound of conf_level\% confidence interval for diamond ratio
\item I2 -  I2 measure of heterogeneity
\item I2_LL - Lower bound of conf_level\% confidence interval for I2
\item I2_UL - upper bound of conf_level\% confidence interval for I2
\item PI_LL - lower bound of conf_level\% of prediction interval
\item PI_UL - upper bound of conf_level\% of prediction interval
\item p - p value for the meta-analytic effect size, based on null of exactly 0
\item width - width of the effect-size confidence interval
\item FE_effect_size - effect size of the fixed-effects model (regardless of if fixed effects was selected
\item RE_effect_size - effect size of the random-effects model (regardless of if random effects was selected
\item FE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio
\item RE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio
}
\item es_heterogeneity - A data frame of of heterogeneity values and
conf_level\% CIs for the meta-analytic effect size.  If a moderator was defined
also reports heterogeneity estimates for each level of the moderator.
\itemize{
\item effect_label - study label
\item moderator_variable_name - if moderator passed, gives name of the moderator
\item moderator_variable_name - 'Overall' and each level of moderator, if passed
\item measure - Name of the measure of heterogeneity
\item LL - lower bound of conf_level\% confidence interval
\item UL - upper bound of conf_level\% confidence interval
}
\item raw_data - A data from with one row for each study that was passed
\itemize{
\item label - study label
\item effect_size - effect size
\item weight - study weight in the meta analysis
\item sample_variance - expected level of sampling variation
\item SE - expected standard error
\item LL - lower bound of conf_level\% confidence interval
\item UL - upper bound of conf_level\% confidence interval
\item mean - used to calculate study p value; this is the d value entered for the study
\item n - study sample size
\item sd - use to calculate study p value; set to 1 for each study
\item p - p value for the study, based on null of exactly 0
}
}
}
\description{
\code{meta_any} is suitable for synthesizing any effect size across
multiple studies.  You must provide the effect size for each study and the
predicted sampling variance for each study.
}
\details{
The meta-analytic effect size, confidence interval and heterogeneity
estimates all come from \code{\link[metafor:rma.uni]{metafor::rma()}}.

The diamond ratio and its confidence interval come from
\code{\link[=CI_diamond_ratio]{CI_diamond_ratio()}}.
}
