% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_d1.R
\name{meta_d1}
\alias{meta_d1}
\title{Estimate ameta-analytic Cohen's d1 across multiple studies}
\usage{
meta_d1(
  data,
  ds,
  ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  random_effects = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{ds}{Set of bias-adjusted cohen's d1 values, 1 for each study}

\item{ns}{Set of sample sizes, positive integers, 1 for each study}

\item{labels}{Optional set of labels, 1 for each study}

\item{moderator}{Optional factor as a categorical moderator; should have k >
2 per group}

\item{contrast}{Optional vector specifying a contrast between moderator
levels}

\item{effect_label}{Optional character providing a human-friendly label for
the effect}

\item{random_effects}{Boolean; TRUE for a random effects model; otherwise
fixed effects}

\item{conf_level}{The confidence level for the confidence interval.  Given in
decimal form.  Defaults to 0.95.}
}
\value{
An esci-estimate object; a list of data frames and properties.
Returned tables include:
\itemize{
\item es_meta - A data frame with one overall.  If a moderator was
defined, there is an additional row for each level of the moderator.
\itemize{
\item effect_label -  Study label
\item effect_size - Effect size
\item LL - Lower bound of conf_level\% confidence interval
\item UL - Upper bound of conf_level\% confidence interval
\item SE - Expected standard error
\item k - Number of studies
\item diamond_ratio - ratio of random to fixed effects meta-analtics effect sizes
\item diamond_ratio_LL - lower bound of conf_level\% confidence interval for diamond ratio
\item diamond_ratio_UL - upper bound of conf_level\% confidence interval for diamond ratio
\item I2 -  I2 measure of heterogeneity
\item I2_LL - Lower bound of conf_level\% confidence interval for I2
\item I2_UL - upper bound of conf_level\% confidence interval for I2
\item PI_LL - lower bound of conf_level\% of prediction interval
\item PI_UL - upper bound of conf_level\% of prediction interval
\item p - p value for the meta-analytic effect size, based on null of exactly 0
\item width - width of the effect-size confidence interval
\item FE_effect_size - effect size of the fixed-effects model (regardless of if fixed effects was selected
\item RE_effect_size - effect size of the random-effects model (regardless of if random effects was selected
\item FE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio
\item RE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio
}
\item es_heterogeneity - A data frame of of heterogeneity values and
conf_level\% CIs for the meta-analytic effect size.  If a moderator was defined
also reports heterogeneity estimates for each level of the moderator.
\itemize{
\item effect_label - study label
\item moderator_variable_name - if moderator passed, gives name of the moderator
\item moderator_variable_name - 'Overall' and each level of moderator, if passed
\item measure - Name of the measure of heterogeneity
\item LL - lower bound of conf_level\% confidence interval
\item UL - upper bound of conf_level\% confidence interval
}
\item raw_data - A data from with one row for each study that was passed
\itemize{
\item label - study label
\item effect_size - effect size
\item weight - study weight in the meta analysis
\item sample_variance - expected level of sampling variation
\item SE - expected standard error
\item LL - lower bound of conf_level\% confidence interval
\item UL - upper bound of conf_level\% confidence interval
\item mean - used to calculate study p value; this is the d value entered for the study
\item n - study sample size
\item sd - use to calculate study p value; set to 1 for each study
\item p - p value for the study, based on null of exactly 0
}
}
}
\description{
\code{meta_d1} is suitable for synthesizing across multiple
single-group studies with a continuous outcome variable, but where the
outcome is not measured on the same scale in all studies
}
\details{
Each study's effect size should be expressed as Cohen's
d1: (mean - reference) / sd.

And the d1 values should all be corrected for bias.
The function \code{\link[=CI_smd_one]{CI_smd_one()}} can assist with converting raw data from
each study to d1_unbiased.

The meta-analytic effect size, confidence interval and heterogeneity
estimates all come from \code{\link[metafor:rma.uni]{metafor::rma()}}.

The diamond ratio and its confidence interval come from
\code{\link[=CI_diamond_ratio]{CI_diamond_ratio()}}.
}
