% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mdiff.R
\name{plot_rdiff}
\alias{plot_rdiff}
\title{Plots for comparing Pearson \emph{r} values between conditions}
\usage{
plot_rdiff(
  estimate,
  error_layout = c("halfeye", "eye", "gradient", "none"),
  error_scale = 0.3,
  error_normalize = c("groups", "all", "panels"),
  difference_axis_breaks = 5,
  simple_contrast_labels = TRUE,
  ylim = c(NA, NA),
  ybreaks = 5,
  rope = c(NA, NA),
  ggtheme = NULL
)
}
\arguments{
\item{estimate}{An esci-estimate object generated by an estimate_pdiff_
function}

\item{error_layout}{Optional; one of 'halfeye', 'eye', 'gradient' or 'none'
to determine how expected error distribution will be displayed for each
estimated parameter.  Defaults to 'halfeye'.  Currently does not apply
if 'median' is selected as effect size, in which case a simple error bar
will be used}

\item{error_scale}{Optional numeric determining width of the expected error
distribution.  Defaults to 0.3}

\item{error_normalize}{Optional; one of 'groups', 'all', or 'panels' to
determine how width of the expected error distributions will be normalized.
Defaults to 'groups'.  See documentation in ggdist}

\item{difference_axis_breaks}{Optional numeric > 1 of suggested number of breaks
for the difference axis.  Defaults to 5}

\item{simple_contrast_labels}{Optional logical to determine if contrasts are
given simple labels ('Reference', 'Comparison', 'Difference') or more
descriptive labels based on the contrast specified.}

\item{ylim}{Optional 2-item vector specifying y-axis limits.  Defaults to
c(NA NA); Use NA to specify auto-limit.}

\item{ybreaks}{Optional numeric > 2 for suggested number of y-axis breaks;
defaults to 5}

\item{rope}{Optional 2-item vector with item 2 >= item 1.  Use to specify a range
of values to use to visualize a hypothesis test.  If both values are the
same, a point-null hypothesis test will be visualized.  If item2 > item1
an interval-null hypothesis test will be visualized.  Defaults to c(NA, NA),
which is to not visualize a hypothesis test}

\item{ggtheme}{Optional ggplot2 theme object to specify the visual style of the
plot.  Defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}}
}
\description{
\code{plot_rdiff} helps visualize comparisons of Pearson's r
estimates between conditions.  It plots the Pearson's r value for each
level of a grouping variable and emphasizes a 1-df comparison among
conditions, plotting the estimated difference and its confidence interval
with a difference axis.  You can pass esci-estimate objects generated
by \code{\link[=estimate_rdiff_two]{estimate_rdiff_two()}}.  This function returns a ggplot2 object.
}
\details{
Use this function to visualize any design analyzed with an estimate_rdiff_
function in esci.
}
