% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generates a scatter plot of data for two continuous variables}
\usage{
plot_scatter(
  estimate,
  show_line = FALSE,
  show_line_CI = FALSE,
  show_PI = FALSE,
  show_residuals = FALSE,
  show_mean_lines = FALSE,
  show_r = FALSE,
  predict_from_x = NULL,
  plot_as_z = FALSE,
  ggtheme = ggplot2::theme_classic()
)
}
\arguments{
\item{estimate}{\itemize{
\item an esci_estimate object generated by \code{\link[=estimate_correlation]{estimate_correlation()}}
}}

\item{show_line}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show the
regression line
}}

\item{show_line_CI}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show the
confidence interval on the regression line
}}

\item{show_PI}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show prediction
intervals
}}

\item{show_residuals}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to show
residuals of prediction
}}

\item{show_mean_lines}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to plot
lines showing the mean of each variable
}}

\item{show_r}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to print the \emph{r}
value and its CI on the plot
}}

\item{predict_from_x}{\itemize{
\item Optional real number in the range of the x variable
for the plot; Defaults to NULL; if passed, the graph shows the predicted
Y' for this x value
}}

\item{plot_as_z}{\itemize{
\item Boolean; defaults to FALSE; set to TRUE to convert
x and y scores to z scores prior to plotting
}}

\item{ggtheme}{\itemize{
\item Optional ggplot2 theme object to control overall styling; defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_classic()}}
}}
}
\description{
\code{plot_scatter} returns a ggplot2 object of data from two continuous variables.  Can
indicate regression line and its confidence interval,prediction intervals
regression residuals and more.
}
\details{
This function requires as input an esci_estimate object generated by
\code{\link[=estimate_correlation]{estimate_correlation()}}
}
