% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_evaluation.R
\name{test_mdiff}
\alias{test_mdiff}
\title{Test a hypothesis about a difference in a continuous outcome variable.}
\usage{
test_mdiff(
  estimate,
  effect_size = c("mean", "median"),
  rope = c(0, 0),
  rope_units = c("raw", "sd"),
  output_html = FALSE
)
}
\arguments{
\item{estimate}{\itemize{
\item An esci_estimate object generated by an
estimate_mdiff_ function
}}

\item{effect_size}{\itemize{
\item One of 'mean' or 'median'.  The effect size selected
must be available in the esci_estimate object; medians are only available
when the estimate was generated from raw data.
}}

\item{rope}{\itemize{
\item A two-element vector defining the Region of Practical
Equivalence (ROPE). Specify c(0, 0) to test a point null of exactly 0.
Specify any two ascending values to test an interval null (e.g. c(-1, 1) to
test the hypothesis tha the difference is between -1 and 1).
}}

\item{rope_units}{\itemize{
\item One of 'raw' (default) or 'sd', specifies the units of
the ROPE. If 'sd' is specified, the rope is defined in standard deviation
units (e.g. c(-1, 1) is taken as between -1 and 1 \emph{standard deviations}
from 0).
}}

\item{output_html}{\itemize{
\item TRUE to return results in HTML; FALSE (default) to return
standard output
}}
}
\description{
\code{test_mdiff} is suitable for conducting a testing a hypothesis about the
magnitude of difference between two conditions for a continuous outcome
variable.  It can test hypotheses about differences in means or medians for
both independent and paired designs.
}
\details{
This function can be passed an esci_estimate object generated by
\code{\link[=estimate_mdiff_one]{estimate_mdiff_one()}}, \code{\link[=estimate_mdiff_two]{estimate_mdiff_two()}},
\code{\link[=estimate_mdiff_paired]{estimate_mdiff_paired()}}, or \code{\link[=estimate_mdiff_ind_contrast]{estimate_mdiff_ind_contrast()}}.

It can test hypotheses about a specific value for the difference (a
point null) or about a range of values (an interval null)
}
