% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_mdiff.R
\name{test_rdiff}
\alias{test_rdiff}
\title{Test a hypothesis about a difference in correlation strength}
\usage{
test_rdiff(estimate, rope = c(0, 0), output_html = FALSE)
}
\arguments{
\item{estimate}{\itemize{
\item An esci_estimate object generated by an estimate_rdiff_
function
}}

\item{rope}{\itemize{
\item A two-element vector defining the Region of Practical
Equivalence (ROPE). Specify c(0, 0) to test a point null of exactly 0.
Specify any two ascending values between -1 and 1 to test an interval null
(e.g. c(-.25, .25) to test the hypothesis that the difference in
correlation is between -.25 and .25).
}}

\item{output_html}{\itemize{
\item TRUE to return results in HTML; FALSE (default) to
return standard output
}}
}
\description{
\code{test_rdiff} is suitable for testing a hypothesis about a
difference in correlation (\emph{r}) between two conditions.
At the moment, it can only test hypotheses for independent-group designs.
}
\details{
This function can be passed an esci_estimate object generated by
\code{\link[=estimate_rdiff_two]{estimate_rdiff_two()}}.

It can test hypotheses about a specific value for the difference (a
point null) or about a range of values (an interval null)
}
